calculate the virtual dataset of a patient

calculate the virtual dataset of a patient

Usage

Virtual_med_one(
  birth_start_date,
  birth_end_date,
  study_start_date,
  study_end_date,
  average_year_visit_n
)

Arguments

birth_start_date

Minimum expected birth date of the patient.

birth_end_date

Maximum expected birth date of the patient.

study_start_date

The starting time of the baseline period specified in the study.

study_end_date

The end time of the follow-up period specified in the study

average_year_visit_n

The average number of visits by patients each year is 6 by default. The data are from the statistical records of the Chinese people.

Value

Returns a data set of a patients, including a doctor's table, prescription table, diagnosis table, and biochemical test table.

Examples

Virtual_med_one(
 birth_start_date = "1980-01-01",birth_end_date ="2000-12-31" ,
   study_start_date = "2014-01-01",study_end_date ="2020-12-31",
   average_year_visit_n = 6)
#> $pat
#>     visit_id sex  birthdate patient_type admission_date Discharge_date
#> 1  leuew3201 men 1981-01-06         Hosp     2014-03-15     2014-03-23
#> 2  nwwjj3122 men 1981-01-06         Hosp     2015-01-25     2015-02-02
#> 3  skapq2552 men 1981-01-06         outp     2015-06-13     2015-06-21
#> 4  qiyfz5616 men 1981-01-06         outp     2015-07-11     2015-07-19
#> 5  qnhvi1738 men 1981-01-06         outp     2015-07-16     2015-07-24
#> 6  hliue1241 men 1981-01-06         Hosp     2016-01-18     2016-01-26
#> 7  arfjs3355 men 1981-01-06         Hosp     2016-04-09     2016-04-17
#> 8  sjdlx4620 men 1981-01-06         Hosp     2017-11-26     2017-12-04
#> 9  nbwts2369 men 1981-01-06         outp     2018-04-06     2018-04-14
#> 10 daffj3622 men 1981-01-06         Hosp     2018-07-01     2018-07-09
#> 11 dalfx4991 men 1981-01-06         outp     2018-09-27     2018-10-05
#> 12 iulxr5057 men 1981-01-06         outp     2018-10-13     2018-10-21
#> 13 pvcac5017 men 1981-01-06         outp     2018-12-16     2018-12-24
#> 14 adopa4779 men 1981-01-06         outp     2019-05-08     2019-05-16
#> 15 vnnop5923 men 1981-01-06         Hosp     2019-05-30     2019-06-07
#> 16 knxfx3127 men 1981-01-06         Hosp     2020-06-06     2020-06-14
#> 17 zlvvg3266 men 1981-01-06         outp     2020-07-15     2020-07-23
#> 18 ejrbv4458 men 1981-01-06         Hosp     2020-09-16     2020-09-24
#> 19 xuflz4883 men 1981-01-06         outp     2020-09-24     2020-10-02
#> 20 paurl1471 men 1981-01-06         outp     2020-09-30     2020-10-08
#> 
#> $prec
#>      visit_id patient_type admission_date              prec_name  prec_date
#> 1   leuew3201         Hosp     2014-03-15              Glinifera 2014-03-16
#> 2   leuew3201         Hosp     2014-03-15         Anticoagulants 2014-03-16
#> 3   leuew3201         Hosp     2014-03-15                   ARBs 2014-03-16
#> 4   leuew3201         Hosp     2014-03-15          beta-blockers 2014-03-16
#> 5   leuew3201         Hosp     2014-03-15 Lipid-regulating drugs 2014-03-16
#> 6   leuew3201         Hosp     2014-03-15                Statins 2014-03-16
#> 7   leuew3201         Hosp     2014-03-15             Biguanides 2014-03-16
#> 8   leuew3201         Hosp     2014-03-15                insulin 2014-03-16
#> 9   leuew3201         Hosp     2014-03-15 antiplatelet medicines 2014-03-16
#> 10  nwwjj3122         Hosp     2015-01-25          beta-blockers 2015-01-26
#> 11  nwwjj3122         Hosp     2015-01-25                   ARBs 2015-01-26
#> 12  nwwjj3122         Hosp     2015-01-25          Sulfonylureas 2015-01-26
#> 13  skapq2552         outp     2015-06-13 antiplatelet medicines 2015-06-14
#> 14  skapq2552         outp     2015-06-13                Statins 2015-06-14
#> 15  skapq2552         outp     2015-06-13          beta-blockers 2015-06-14
#> 16  skapq2552         outp     2015-06-13                   ARBs 2015-06-14
#> 17  skapq2552         outp     2015-06-13              Glinifera 2015-06-14
#> 18  skapq2552         outp     2015-06-13 Lipid-regulating drugs 2015-06-14
#> 19  skapq2552         outp     2015-06-13          Sulfonylureas 2015-06-14
#> 20  qiyfz5616         outp     2015-07-11                   ARBs 2015-07-12
#> 21  qiyfz5616         outp     2015-07-11                insulin 2015-07-12
#> 22  qiyfz5616         outp     2015-07-11              Glinifera 2015-07-12
#> 23  qiyfz5616         outp     2015-07-11 Lipid-regulating drugs 2015-07-12
#> 24  qiyfz5616         outp     2015-07-11          Sulfonylureas 2015-07-12
#> 25  qiyfz5616         outp     2015-07-11         Anticoagulants 2015-07-12
#> 26  qnhvi1738         outp     2015-07-16          beta-blockers 2015-07-17
#> 27  qnhvi1738         outp     2015-07-16 Lipid-regulating drugs 2015-07-17
#> 28  qnhvi1738         outp     2015-07-16          Sulfonylureas 2015-07-17
#> 29  hliue1241         Hosp     2016-01-18          Sulfonylureas 2016-01-19
#> 30  hliue1241         Hosp     2016-01-18             Biguanides 2016-01-19
#> 31  hliue1241         Hosp     2016-01-18              Glinifera 2016-01-19
#> 32  hliue1241         Hosp     2016-01-18          beta-blockers 2016-01-19
#> 33  hliue1241         Hosp     2016-01-18         Anticoagulants 2016-01-19
#> 34  hliue1241         Hosp     2016-01-18                Statins 2016-01-19
#> 35  hliue1241         Hosp     2016-01-18                   ARBs 2016-01-19
#> 36  hliue1241         Hosp     2016-01-18 antiplatelet medicines 2016-01-19
#> 37  hliue1241         Hosp     2016-01-18 Lipid-regulating drugs 2016-01-19
#> 38  arfjs3355         Hosp     2016-04-09 antiplatelet medicines 2016-04-10
#> 39  arfjs3355         Hosp     2016-04-09 Lipid-regulating drugs 2016-04-10
#> 40  arfjs3355         Hosp     2016-04-09                   ARBs 2016-04-10
#> 41  arfjs3355         Hosp     2016-04-09                insulin 2016-04-10
#> 42  arfjs3355         Hosp     2016-04-09          Sulfonylureas 2016-04-10
#> 43  arfjs3355         Hosp     2016-04-09          beta-blockers 2016-04-10
#> 44  arfjs3355         Hosp     2016-04-09              Glinifera 2016-04-10
#> 45  sjdlx4620         Hosp     2017-11-26         Anticoagulants 2017-11-27
#> 46  sjdlx4620         Hosp     2017-11-26             Biguanides 2017-11-27
#> 47  sjdlx4620         Hosp     2017-11-26 antiplatelet medicines 2017-11-27
#> 48  nbwts2369         outp     2018-04-06 antiplatelet medicines 2018-04-07
#> 49  nbwts2369         outp     2018-04-06              Glinifera 2018-04-07
#> 50  nbwts2369         outp     2018-04-06 Lipid-regulating drugs 2018-04-07
#> 51  daffj3622         Hosp     2018-07-01 Lipid-regulating drugs 2018-07-02
#> 52  daffj3622         Hosp     2018-07-01                Statins 2018-07-02
#> 53  daffj3622         Hosp     2018-07-01          beta-blockers 2018-07-02
#> 54  daffj3622         Hosp     2018-07-01                insulin 2018-07-02
#> 55  daffj3622         Hosp     2018-07-01 antiplatelet medicines 2018-07-02
#> 56  daffj3622         Hosp     2018-07-01              Glinifera 2018-07-02
#> 57  dalfx4991         outp     2018-09-27          Sulfonylureas 2018-09-28
#> 58  dalfx4991         outp     2018-09-27          beta-blockers 2018-09-28
#> 59  dalfx4991         outp     2018-09-27                insulin 2018-09-28
#> 60  dalfx4991         outp     2018-09-27 antiplatelet medicines 2018-09-28
#> 61  dalfx4991         outp     2018-09-27                Statins 2018-09-28
#> 62  dalfx4991         outp     2018-09-27             Biguanides 2018-09-28
#> 63  dalfx4991         outp     2018-09-27 Lipid-regulating drugs 2018-09-28
#> 64  dalfx4991         outp     2018-09-27         Anticoagulants 2018-09-28
#> 65  iulxr5057         outp     2018-10-13         Anticoagulants 2018-10-14
#> 66  iulxr5057         outp     2018-10-13 Lipid-regulating drugs 2018-10-14
#> 67  iulxr5057         outp     2018-10-13              Glinifera 2018-10-14
#> 68  iulxr5057         outp     2018-10-13          beta-blockers 2018-10-14
#> 69  iulxr5057         outp     2018-10-13          Sulfonylureas 2018-10-14
#> 70  iulxr5057         outp     2018-10-13                Statins 2018-10-14
#> 71  iulxr5057         outp     2018-10-13                insulin 2018-10-14
#> 72  iulxr5057         outp     2018-10-13                   ARBs 2018-10-14
#> 73  iulxr5057         outp     2018-10-13 antiplatelet medicines 2018-10-14
#> 74  pvcac5017         outp     2018-12-16              Glinifera 2018-12-17
#> 75  pvcac5017         outp     2018-12-16 Lipid-regulating drugs 2018-12-17
#> 76  pvcac5017         outp     2018-12-16                   ARBs 2018-12-17
#> 77  pvcac5017         outp     2018-12-16                insulin 2018-12-17
#> 78  pvcac5017         outp     2018-12-16          Sulfonylureas 2018-12-17
#> 79  pvcac5017         outp     2018-12-16 antiplatelet medicines 2018-12-17
#> 80  pvcac5017         outp     2018-12-16             Biguanides 2018-12-17
#> 81  adopa4779         outp     2019-05-08                Statins 2019-05-09
#> 82  adopa4779         outp     2019-05-08             Biguanides 2019-05-09
#> 83  adopa4779         outp     2019-05-08                   ARBs 2019-05-09
#> 84  vnnop5923         Hosp     2019-05-30                insulin 2019-05-31
#> 85  vnnop5923         Hosp     2019-05-30                   ARBs 2019-05-31
#> 86  vnnop5923         Hosp     2019-05-30 Lipid-regulating drugs 2019-05-31
#> 87  knxfx3127         Hosp     2020-06-06             Biguanides 2020-06-07
#> 88  knxfx3127         Hosp     2020-06-06                insulin 2020-06-07
#> 89  knxfx3127         Hosp     2020-06-06 antiplatelet medicines 2020-06-07
#> 90  knxfx3127         Hosp     2020-06-06                   ARBs 2020-06-07
#> 91  knxfx3127         Hosp     2020-06-06         Anticoagulants 2020-06-07
#> 92  zlvvg3266         outp     2020-07-15                Statins 2020-07-16
#> 93  zlvvg3266         outp     2020-07-15                insulin 2020-07-16
#> 94  zlvvg3266         outp     2020-07-15             Biguanides 2020-07-16
#> 95  zlvvg3266         outp     2020-07-15 Lipid-regulating drugs 2020-07-16
#> 96  zlvvg3266         outp     2020-07-15 antiplatelet medicines 2020-07-16
#> 97  ejrbv4458         Hosp     2020-09-16                insulin 2020-09-17
#> 98  ejrbv4458         Hosp     2020-09-16                   ARBs 2020-09-17
#> 99  ejrbv4458         Hosp     2020-09-16 Lipid-regulating drugs 2020-09-17
#> 100 ejrbv4458         Hosp     2020-09-16             Biguanides 2020-09-17
#> 101 xuflz4883         outp     2020-09-24                   ARBs 2020-09-25
#> 102 xuflz4883         outp     2020-09-24         Anticoagulants 2020-09-25
#> 103 xuflz4883         outp     2020-09-24          Sulfonylureas 2020-09-25
#> 104 xuflz4883         outp     2020-09-24                Statins 2020-09-25
#> 105 xuflz4883         outp     2020-09-24 antiplatelet medicines 2020-09-25
#> 106 paurl1471         outp     2020-09-30                   ARBs 2020-10-01
#> 107 paurl1471         outp     2020-09-30         Anticoagulants 2020-10-01
#> 108 paurl1471         outp     2020-09-30                Statins 2020-10-01
#> 109 paurl1471         outp     2020-09-30              Glinifera 2020-10-01
#> 110 paurl1471         outp     2020-09-30             Biguanides 2020-10-01
#> 111 paurl1471         outp     2020-09-30 antiplatelet medicines 2020-10-01
#> 
#> $lab
#>     visit_id patient_type admission_date lab_name lab_va  test_date
#> 1  leuew3201         Hosp     2014-03-15      LDL      3 2014-03-15
#> 2  leuew3201         Hosp     2014-03-15      WBC     12 2014-03-15
#> 3  leuew3201         Hosp     2014-03-15       TC      3 2014-03-15
#> 4  leuew3201         Hosp     2014-03-15      HDL      2 2014-03-15
#> 5  nwwjj3122         Hosp     2015-01-25    HbA1c      9 2015-01-25
#> 6  nwwjj3122         Hosp     2015-01-25      HDL      5 2015-01-25
#> 7  nwwjj3122         Hosp     2015-01-25      LDL      2 2015-01-25
#> 8  nwwjj3122         Hosp     2015-01-25       TG     18 2015-01-25
#> 9  skapq2552         outp     2015-06-13       TG     11 2015-06-13
#> 10 skapq2552         outp     2015-06-13      WBC     13 2015-06-13
#> 11 skapq2552         outp     2015-06-13    HbA1c      7 2015-06-13
#> 12 skapq2552         outp     2015-06-13      LDL      3 2015-06-13
#> 13 skapq2552         outp     2015-06-13       TC      1 2015-06-13
#> 14 skapq2552         outp     2015-06-13      HDL      2 2015-06-13
#> 15 qiyfz5616         outp     2015-07-11    HbA1c      4 2015-07-11
#> 16 qiyfz5616         outp     2015-07-11       TG      4 2015-07-11
#> 17 qiyfz5616         outp     2015-07-11      HDL      1 2015-07-11
#> 18 qiyfz5616         outp     2015-07-11      WBC      7 2015-07-11
#> 19 qnhvi1738         outp     2015-07-16      HDL      4 2015-07-16
#> 20 qnhvi1738         outp     2015-07-16       TG     13 2015-07-16
#> 21 qnhvi1738         outp     2015-07-16    HbA1c      8 2015-07-16
#> 22 qnhvi1738         outp     2015-07-16      LDL      4 2015-07-16
#> 23 qnhvi1738         outp     2015-07-16      WBC     14 2015-07-16
#> 24 qnhvi1738         outp     2015-07-16       TC      4 2015-07-16
#> 25 hliue1241         Hosp     2016-01-18       TC      6 2016-01-18
#> 26 hliue1241         Hosp     2016-01-18      WBC      3 2016-01-18
#> 27 hliue1241         Hosp     2016-01-18       TG      1 2016-01-18
#> 28 hliue1241         Hosp     2016-01-18      HDL      3 2016-01-18
#> 29 hliue1241         Hosp     2016-01-18    HbA1c      4 2016-01-18
#> 30 hliue1241         Hosp     2016-01-18      LDL      4 2016-01-18
#> 31 arfjs3355         Hosp     2016-04-09       TG     15 2016-04-09
#> 32 arfjs3355         Hosp     2016-04-09    HbA1c      3 2016-04-09
#> 33 arfjs3355         Hosp     2016-04-09      LDL      4 2016-04-09
#> 34 sjdlx4620         Hosp     2017-11-26       TG      3 2017-11-26
#> 35 sjdlx4620         Hosp     2017-11-26    HbA1c     12 2017-11-26
#> 36 sjdlx4620         Hosp     2017-11-26      LDL      1 2017-11-26
#> 37 sjdlx4620         Hosp     2017-11-26       TC      2 2017-11-26
#> 38 nbwts2369         outp     2018-04-06       TC      2 2018-04-06
#> 39 nbwts2369         outp     2018-04-06      LDL      5 2018-04-06
#> 40 nbwts2369         outp     2018-04-06       TG     14 2018-04-06
#> 41 nbwts2369         outp     2018-04-06    HbA1c      3 2018-04-06
#> 42 daffj3622         Hosp     2018-07-01      LDL      4 2018-07-01
#> 43 daffj3622         Hosp     2018-07-01    HbA1c     12 2018-07-01
#> 44 daffj3622         Hosp     2018-07-01      WBC     11 2018-07-01
#> 45 daffj3622         Hosp     2018-07-01      HDL      2 2018-07-01
#> 46 daffj3622         Hosp     2018-07-01       TC      4 2018-07-01
#> 47 daffj3622         Hosp     2018-07-01       TG      9 2018-07-01
#> 48 dalfx4991         outp     2018-09-27      LDL      2 2018-09-27
#> 49 dalfx4991         outp     2018-09-27      HDL      1 2018-09-27
#> 50 dalfx4991         outp     2018-09-27       TG      6 2018-09-27
#> 51 dalfx4991         outp     2018-09-27      WBC      7 2018-09-27
#> 52 dalfx4991         outp     2018-09-27    HbA1c      4 2018-09-27
#> 53 iulxr5057         outp     2018-10-13       TG     20 2018-10-13
#> 54 iulxr5057         outp     2018-10-13      HDL      3 2018-10-13
#> 55 iulxr5057         outp     2018-10-13      WBC     10 2018-10-13
#> 56 iulxr5057         outp     2018-10-13      LDL      2 2018-10-13
#> 57 iulxr5057         outp     2018-10-13    HbA1c      8 2018-10-13
#> 58 iulxr5057         outp     2018-10-13       TC      6 2018-10-13
#> 59 pvcac5017         outp     2018-12-16    HbA1c      3 2018-12-16
#> 60 pvcac5017         outp     2018-12-16      WBC     19 2018-12-16
#> 61 pvcac5017         outp     2018-12-16       TG      3 2018-12-16
#> 62 pvcac5017         outp     2018-12-16       TC      4 2018-12-16
#> 63 pvcac5017         outp     2018-12-16      LDL      1 2018-12-16
#> 64 pvcac5017         outp     2018-12-16      HDL      3 2018-12-16
#> 65 adopa4779         outp     2019-05-08      LDL      2 2019-05-08
#> 66 adopa4779         outp     2019-05-08       TC      4 2019-05-08
#> 67 adopa4779         outp     2019-05-08      HDL      2 2019-05-08
#> 68 adopa4779         outp     2019-05-08    HbA1c      2 2019-05-08
#> 69 adopa4779         outp     2019-05-08       TG     18 2019-05-08
#> 70 adopa4779         outp     2019-05-08      WBC      9 2019-05-08
#> 71 vnnop5923         Hosp     2019-05-30    HbA1c      7 2019-05-30
#> 72 vnnop5923         Hosp     2019-05-30      WBC     18 2019-05-30
#> 73 vnnop5923         Hosp     2019-05-30       TG      3 2019-05-30
#> 74 vnnop5923         Hosp     2019-05-30       TC      7 2019-05-30
#> 75 vnnop5923         Hosp     2019-05-30      LDL      5 2019-05-30
#> 76 vnnop5923         Hosp     2019-05-30      HDL      5 2019-05-30
#> 77 knxfx3127         Hosp     2020-06-06      HDL      2 2020-06-06
#> 78 knxfx3127         Hosp     2020-06-06      LDL      5 2020-06-06
#> 79 knxfx3127         Hosp     2020-06-06      WBC      4 2020-06-06
#> 80 knxfx3127         Hosp     2020-06-06    HbA1c     10 2020-06-06
#> 81 zlvvg3266         outp     2020-07-15      HDL      2 2020-07-15
#> 82 zlvvg3266         outp     2020-07-15    HbA1c     12 2020-07-15
#> 83 zlvvg3266         outp     2020-07-15      WBC     19 2020-07-15
#> 84 ejrbv4458         Hosp     2020-09-16      LDL      3 2020-09-16
#> 85 ejrbv4458         Hosp     2020-09-16       TC      5 2020-09-16
#> 86 ejrbv4458         Hosp     2020-09-16       TG      9 2020-09-16
#> 87 ejrbv4458         Hosp     2020-09-16      HDL      2 2020-09-16
#> 88 ejrbv4458         Hosp     2020-09-16    HbA1c      7 2020-09-16
#> 89 xuflz4883         outp     2020-09-24       TC      7 2020-09-24
#> 90 xuflz4883         outp     2020-09-24      LDL      4 2020-09-24
#> 91 xuflz4883         outp     2020-09-24      WBC     15 2020-09-24
#> 92 xuflz4883         outp     2020-09-24       TG      7 2020-09-24
#> 93 xuflz4883         outp     2020-09-24      HDL      2 2020-09-24
#> 94 paurl1471         outp     2020-09-30       TC      3 2020-09-30
#> 95 paurl1471         outp     2020-09-30      WBC     17 2020-09-30
#> 96 paurl1471         outp     2020-09-30      HDL      2 2020-09-30
#> 97 paurl1471         outp     2020-09-30      LDL      3 2020-09-30
#> 
#> $diag
#> # A tibble: 20 × 5
#>    visit_id  patient_type admission_date diag_date  diag_union                  
#>    <chr>     <chr>        <date>         <date>     <chr>                       
#>  1 leuew3201 Hosp         2014-03-15     2014-03-16 high blood pressure,diabete…
#>  2 nwwjj3122 Hosp         2015-01-25     2015-01-26 diabetes,NA,Stroke,Symptoma…
#>  3 skapq2552 outp         2015-06-13     2015-06-14 Chronic lung disease,Peptic…
#>  4 qiyfz5616 outp         2015-07-11     2015-07-12 Mild liver disease,NA,diabe…
#>  5 qnhvi1738 outp         2015-07-16     2015-07-17 NA,Dyslipidemia,Peptic ulce…
#>  6 hliue1241 Hosp         2016-01-18     2016-01-19 Lymphatic carcinoma,Dyslipi…
#>  7 arfjs3355 Hosp         2016-04-09     2016-04-10 NA,Stroke,Dementia,Connecti…
#>  8 sjdlx4620 Hosp         2017-11-26     2017-11-27 NA                          
#>  9 nbwts2369 outp         2018-04-06     2018-04-07 NA                          
#> 10 daffj3622 Hosp         2018-07-01     2018-07-02 NA                          
#> 11 dalfx4991 outp         2018-09-27     2018-09-28 NA,Peripheral vascular dise…
#> 12 iulxr5057 outp         2018-10-13     2018-10-14 Lymphatic carcinoma,high bl…
#> 13 pvcac5017 outp         2018-12-16     2018-12-17 Lymphatic carcinoma,NA,Chro…
#> 14 adopa4779 outp         2019-05-08     2019-05-09 NA                          
#> 15 vnnop5923 Hosp         2019-05-30     2019-05-31 NA,high blood pressure,Peri…
#> 16 knxfx3127 Hosp         2020-06-06     2020-06-07 Mild liver disease,Stroke,N…
#> 17 zlvvg3266 outp         2020-07-15     2020-07-16 Dyslipidemia,diabetes,Perip…
#> 18 ejrbv4458 Hosp         2020-09-16     2020-09-17 diabetes,NA,Dementia        
#> 19 xuflz4883 outp         2020-09-24     2020-09-25 Connective tissue disease,d…
#> 20 paurl1471 outp         2020-09-30     2020-10-01 NA,Symptomatic neuropathy,M…
#>