calculate the virtual dataset of a patient

calculate the virtual dataset of a patient

Usage

Virtual_med_one(
  birth_start_date,
  birth_end_date,
  study_start_date,
  study_end_date,
  average_year_visit_n
)

Arguments

birth_start_date

Minimum expected birth date of the patient.

birth_end_date

Maximum expected birth date of the patient.

study_start_date

The starting time of the baseline period specified in the study.

study_end_date

The end time of the follow-up period specified in the study

average_year_visit_n

The average number of visits by patients each year is 6 by default. The data are from the statistical records of the Chinese people.

Value

Returns a data set of a patients, including a doctor's table, prescription table, diagnosis table, and biochemical test table.

Examples

Virtual_med_one(
 birth_start_date = "1980-01-01",birth_end_date ="2000-12-31" ,
   study_start_date = "2014-01-01",study_end_date ="2020-12-31",
   average_year_visit_n = 6)
#> $pat
#>     visit_id sex  birthdate patient_type admission_date Discharge_date
#> 1  gxioj2964 men 1980-11-09         Hosp     2014-11-20     2014-12-19
#> 2  dbjmj5676 men 1980-11-09         Hosp     2014-12-09     2015-01-07
#> 3  cqaix5066 men 1980-11-09         Hosp     2014-12-23     2015-01-21
#> 4  ddzxl1821 men 1980-11-09         outp     2014-12-24     2015-01-22
#> 5  wkmzj1037 men 1980-11-09         outp     2015-01-13     2015-02-11
#> 6  wqudj3630 men 1980-11-09         outp     2015-01-16     2015-02-14
#> 7  rugrj5988 men 1980-11-09         outp     2015-02-06     2015-03-07
#> 8  oezbz4107 men 1980-11-09         Hosp     2015-02-15     2015-03-16
#> 9  keksr5885 men 1980-11-09         outp     2015-04-16     2015-05-15
#> 10 auphy4493 men 1980-11-09         outp     2015-05-01     2015-05-30
#> 11 zfmwv1799 men 1980-11-09         outp     2015-05-19     2015-06-17
#> 12 jqogc2293 men 1980-11-09         Hosp     2015-05-23     2015-06-21
#> 13 cgzet3758 men 1980-11-09         outp     2015-06-21     2015-07-20
#> 14 nwdni4370 men 1980-11-09         outp     2015-08-18     2015-09-16
#> 15 inlmf5251 men 1980-11-09         Hosp     2015-11-03     2015-12-02
#> 16 kykdz3468 men 1980-11-09         Hosp     2016-01-01     2016-01-30
#> 17 pjdin1619 men 1980-11-09         Hosp     2016-04-24     2016-05-23
#> 18 dbxdg2557 men 1980-11-09         outp     2016-08-23     2016-09-21
#> 19 mdpnk4183 men 1980-11-09         outp     2016-09-08     2016-10-07
#> 20 wzcdr3096 men 1980-11-09         outp     2017-01-04     2017-02-02
#> 21 jexpv3996 men 1980-11-09         Hosp     2017-01-29     2017-02-27
#> 22 gzqth4176 men 1980-11-09         Hosp     2017-03-15     2017-04-13
#> 23 ufrzf5777 men 1980-11-09         Hosp     2017-09-18     2017-10-17
#> 24 cxpvu1806 men 1980-11-09         Hosp     2018-06-13     2018-07-12
#> 25 iikmu1400 men 1980-11-09         outp     2018-08-07     2018-09-05
#> 26 yhalt3290 men 1980-11-09         Hosp     2018-08-21     2018-09-19
#> 27 wpcud1752 men 1980-11-09         Hosp     2018-10-02     2018-10-31
#> 28 urnuv5242 men 1980-11-09         Hosp     2018-10-06     2018-11-04
#> 29 wnvqe5313 men 1980-11-09         Hosp     2018-10-16     2018-11-14
#> 30 yqbkk1784 men 1980-11-09         outp     2019-01-05     2019-02-03
#> 31 jtfzn4085 men 1980-11-09         Hosp     2019-07-25     2019-08-23
#> 32 sjkwd4201 men 1980-11-09         Hosp     2019-08-09     2019-09-07
#> 33 pnvsp5269 men 1980-11-09         Hosp     2019-08-11     2019-09-09
#> 34 ngmep2107 men 1980-11-09         outp     2020-02-05     2020-03-05
#> 35 lxlrq2954 men 1980-11-09         outp     2020-02-26     2020-03-26
#> 36 tjhep3934 men 1980-11-09         outp     2020-05-08     2020-06-06
#> 37 dxagx3662 men 1980-11-09         outp     2020-06-24     2020-07-23
#> 38 svzdx2060 men 1980-11-09         outp     2020-07-27     2020-08-25
#> 39 nitlg1679 men 1980-11-09         Hosp     2020-08-17     2020-09-15
#> 
#> $prec
#>      visit_id patient_type admission_date              prec_name  prec_date
#> 1   gxioj2964         Hosp     2014-11-20          Sulfonylureas 2014-11-20
#> 2   gxioj2964         Hosp     2014-11-20              Glinifera 2014-11-20
#> 3   gxioj2964         Hosp     2014-11-20                   ARBs 2014-11-20
#> 4   gxioj2964         Hosp     2014-11-20          beta-blockers 2014-11-20
#> 5   gxioj2964         Hosp     2014-11-20             Biguanides 2014-11-20
#> 6   gxioj2964         Hosp     2014-11-20         Anticoagulants 2014-11-20
#> 7   dbjmj5676         Hosp     2014-12-09 antiplatelet medicines 2014-12-09
#> 8   dbjmj5676         Hosp     2014-12-09                Statins 2014-12-09
#> 9   dbjmj5676         Hosp     2014-12-09              Glinifera 2014-12-09
#> 10  dbjmj5676         Hosp     2014-12-09          beta-blockers 2014-12-09
#> 11  cqaix5066         Hosp     2014-12-23 Lipid-regulating drugs 2014-12-23
#> 12  cqaix5066         Hosp     2014-12-23          beta-blockers 2014-12-23
#> 13  cqaix5066         Hosp     2014-12-23 antiplatelet medicines 2014-12-23
#> 14  cqaix5066         Hosp     2014-12-23                   ARBs 2014-12-23
#> 15  cqaix5066         Hosp     2014-12-23         Anticoagulants 2014-12-23
#> 16  cqaix5066         Hosp     2014-12-23                Statins 2014-12-23
#> 17  cqaix5066         Hosp     2014-12-23              Glinifera 2014-12-23
#> 18  cqaix5066         Hosp     2014-12-23                insulin 2014-12-23
#> 19  cqaix5066         Hosp     2014-12-23          Sulfonylureas 2014-12-23
#> 20  ddzxl1821         outp     2014-12-24 Lipid-regulating drugs 2014-12-24
#> 21  ddzxl1821         outp     2014-12-24             Biguanides 2014-12-24
#> 22  ddzxl1821         outp     2014-12-24         Anticoagulants 2014-12-24
#> 23  wkmzj1037         outp     2015-01-13                Statins 2015-01-13
#> 24  wkmzj1037         outp     2015-01-13          Sulfonylureas 2015-01-13
#> 25  wkmzj1037         outp     2015-01-13                   ARBs 2015-01-13
#> 26  wkmzj1037         outp     2015-01-13          beta-blockers 2015-01-13
#> 27  wqudj3630         outp     2015-01-16              Glinifera 2015-01-16
#> 28  wqudj3630         outp     2015-01-16                Statins 2015-01-16
#> 29  wqudj3630         outp     2015-01-16          beta-blockers 2015-01-16
#> 30  wqudj3630         outp     2015-01-16          Sulfonylureas 2015-01-16
#> 31  wqudj3630         outp     2015-01-16 Lipid-regulating drugs 2015-01-16
#> 32  wqudj3630         outp     2015-01-16             Biguanides 2015-01-16
#> 33  wqudj3630         outp     2015-01-16                insulin 2015-01-16
#> 34  rugrj5988         outp     2015-02-06         Anticoagulants 2015-02-06
#> 35  rugrj5988         outp     2015-02-06              Glinifera 2015-02-06
#> 36  rugrj5988         outp     2015-02-06                   ARBs 2015-02-06
#> 37  rugrj5988         outp     2015-02-06             Biguanides 2015-02-06
#> 38  rugrj5988         outp     2015-02-06          Sulfonylureas 2015-02-06
#> 39  rugrj5988         outp     2015-02-06                insulin 2015-02-06
#> 40  rugrj5988         outp     2015-02-06 antiplatelet medicines 2015-02-06
#> 41  rugrj5988         outp     2015-02-06          beta-blockers 2015-02-06
#> 42  rugrj5988         outp     2015-02-06                Statins 2015-02-06
#> 43  oezbz4107         Hosp     2015-02-15         Anticoagulants 2015-02-15
#> 44  oezbz4107         Hosp     2015-02-15                insulin 2015-02-15
#> 45  oezbz4107         Hosp     2015-02-15             Biguanides 2015-02-15
#> 46  oezbz4107         Hosp     2015-02-15          beta-blockers 2015-02-15
#> 47  oezbz4107         Hosp     2015-02-15                Statins 2015-02-15
#> 48  oezbz4107         Hosp     2015-02-15 antiplatelet medicines 2015-02-15
#> 49  oezbz4107         Hosp     2015-02-15              Glinifera 2015-02-15
#> 50  keksr5885         outp     2015-04-16 antiplatelet medicines 2015-04-16
#> 51  keksr5885         outp     2015-04-16              Glinifera 2015-04-16
#> 52  keksr5885         outp     2015-04-16         Anticoagulants 2015-04-16
#> 53  keksr5885         outp     2015-04-16          Sulfonylureas 2015-04-16
#> 54  keksr5885         outp     2015-04-16                   ARBs 2015-04-16
#> 55  keksr5885         outp     2015-04-16 Lipid-regulating drugs 2015-04-16
#> 56  auphy4493         outp     2015-05-01         Anticoagulants 2015-05-01
#> 57  auphy4493         outp     2015-05-01                   ARBs 2015-05-01
#> 58  auphy4493         outp     2015-05-01                insulin 2015-05-01
#> 59  auphy4493         outp     2015-05-01              Glinifera 2015-05-01
#> 60  auphy4493         outp     2015-05-01 Lipid-regulating drugs 2015-05-01
#> 61  auphy4493         outp     2015-05-01             Biguanides 2015-05-01
#> 62  zfmwv1799         outp     2015-05-19                insulin 2015-05-19
#> 63  zfmwv1799         outp     2015-05-19 Lipid-regulating drugs 2015-05-19
#> 64  zfmwv1799         outp     2015-05-19          beta-blockers 2015-05-19
#> 65  zfmwv1799         outp     2015-05-19              Glinifera 2015-05-19
#> 66  zfmwv1799         outp     2015-05-19                   ARBs 2015-05-19
#> 67  zfmwv1799         outp     2015-05-19          Sulfonylureas 2015-05-19
#> 68  zfmwv1799         outp     2015-05-19 antiplatelet medicines 2015-05-19
#> 69  zfmwv1799         outp     2015-05-19                Statins 2015-05-19
#> 70  zfmwv1799         outp     2015-05-19             Biguanides 2015-05-19
#> 71  jqogc2293         Hosp     2015-05-23                   ARBs 2015-05-23
#> 72  jqogc2293         Hosp     2015-05-23              Glinifera 2015-05-23
#> 73  jqogc2293         Hosp     2015-05-23 antiplatelet medicines 2015-05-23
#> 74  jqogc2293         Hosp     2015-05-23          Sulfonylureas 2015-05-23
#> 75  jqogc2293         Hosp     2015-05-23                insulin 2015-05-23
#> 76  jqogc2293         Hosp     2015-05-23         Anticoagulants 2015-05-23
#> 77  jqogc2293         Hosp     2015-05-23             Biguanides 2015-05-23
#> 78  jqogc2293         Hosp     2015-05-23          beta-blockers 2015-05-23
#> 79  jqogc2293         Hosp     2015-05-23                Statins 2015-05-23
#> 80  cgzet3758         outp     2015-06-21 antiplatelet medicines 2015-06-21
#> 81  cgzet3758         outp     2015-06-21          Sulfonylureas 2015-06-21
#> 82  cgzet3758         outp     2015-06-21                insulin 2015-06-21
#> 83  cgzet3758         outp     2015-06-21             Biguanides 2015-06-21
#> 84  cgzet3758         outp     2015-06-21              Glinifera 2015-06-21
#> 85  cgzet3758         outp     2015-06-21                   ARBs 2015-06-21
#> 86  cgzet3758         outp     2015-06-21         Anticoagulants 2015-06-21
#> 87  cgzet3758         outp     2015-06-21                Statins 2015-06-21
#> 88  nwdni4370         outp     2015-08-18                Statins 2015-08-18
#> 89  nwdni4370         outp     2015-08-18 Lipid-regulating drugs 2015-08-18
#> 90  nwdni4370         outp     2015-08-18         Anticoagulants 2015-08-18
#> 91  nwdni4370         outp     2015-08-18          Sulfonylureas 2015-08-18
#> 92  nwdni4370         outp     2015-08-18              Glinifera 2015-08-18
#> 93  nwdni4370         outp     2015-08-18                   ARBs 2015-08-18
#> 94  nwdni4370         outp     2015-08-18          beta-blockers 2015-08-18
#> 95  nwdni4370         outp     2015-08-18                insulin 2015-08-18
#> 96  nwdni4370         outp     2015-08-18 antiplatelet medicines 2015-08-18
#> 97  inlmf5251         Hosp     2015-11-03          Sulfonylureas 2015-11-03
#> 98  inlmf5251         Hosp     2015-11-03              Glinifera 2015-11-03
#> 99  inlmf5251         Hosp     2015-11-03          beta-blockers 2015-11-03
#> 100 inlmf5251         Hosp     2015-11-03                   ARBs 2015-11-03
#> 101 inlmf5251         Hosp     2015-11-03             Biguanides 2015-11-03
#> 102 kykdz3468         Hosp     2016-01-01 Lipid-regulating drugs 2016-01-01
#> 103 kykdz3468         Hosp     2016-01-01         Anticoagulants 2016-01-01
#> 104 kykdz3468         Hosp     2016-01-01          Sulfonylureas 2016-01-01
#> 105 kykdz3468         Hosp     2016-01-01                Statins 2016-01-01
#> 106 kykdz3468         Hosp     2016-01-01             Biguanides 2016-01-01
#> 107 kykdz3468         Hosp     2016-01-01              Glinifera 2016-01-01
#> 108 kykdz3468         Hosp     2016-01-01                   ARBs 2016-01-01
#> 109 pjdin1619         Hosp     2016-04-24              Glinifera 2016-04-24
#> 110 pjdin1619         Hosp     2016-04-24             Biguanides 2016-04-24
#> 111 pjdin1619         Hosp     2016-04-24          Sulfonylureas 2016-04-24
#> 112 pjdin1619         Hosp     2016-04-24                   ARBs 2016-04-24
#> 113 pjdin1619         Hosp     2016-04-24                insulin 2016-04-24
#> 114 pjdin1619         Hosp     2016-04-24 Lipid-regulating drugs 2016-04-24
#> 115 dbxdg2557         outp     2016-08-23              Glinifera 2016-08-23
#> 116 dbxdg2557         outp     2016-08-23          Sulfonylureas 2016-08-23
#> 117 dbxdg2557         outp     2016-08-23                   ARBs 2016-08-23
#> 118 dbxdg2557         outp     2016-08-23             Biguanides 2016-08-23
#> 119 dbxdg2557         outp     2016-08-23 antiplatelet medicines 2016-08-23
#> 120 dbxdg2557         outp     2016-08-23 Lipid-regulating drugs 2016-08-23
#> 121 dbxdg2557         outp     2016-08-23          beta-blockers 2016-08-23
#> 122 dbxdg2557         outp     2016-08-23         Anticoagulants 2016-08-23
#> 123 mdpnk4183         outp     2016-09-08             Biguanides 2016-09-08
#> 124 mdpnk4183         outp     2016-09-08         Anticoagulants 2016-09-08
#> 125 mdpnk4183         outp     2016-09-08                insulin 2016-09-08
#> 126 mdpnk4183         outp     2016-09-08 Lipid-regulating drugs 2016-09-08
#> 127 mdpnk4183         outp     2016-09-08                   ARBs 2016-09-08
#> 128 mdpnk4183         outp     2016-09-08              Glinifera 2016-09-08
#> 129 wzcdr3096         outp     2017-01-04          Sulfonylureas 2017-01-04
#> 130 wzcdr3096         outp     2017-01-04                   ARBs 2017-01-04
#> 131 wzcdr3096         outp     2017-01-04 antiplatelet medicines 2017-01-04
#> 132 wzcdr3096         outp     2017-01-04                Statins 2017-01-04
#> 133 wzcdr3096         outp     2017-01-04                insulin 2017-01-04
#> 134 wzcdr3096         outp     2017-01-04              Glinifera 2017-01-04
#> 135 wzcdr3096         outp     2017-01-04             Biguanides 2017-01-04
#> 136 wzcdr3096         outp     2017-01-04          beta-blockers 2017-01-04
#> 137 wzcdr3096         outp     2017-01-04 Lipid-regulating drugs 2017-01-04
#> 138 jexpv3996         Hosp     2017-01-29                Statins 2017-01-29
#> 139 jexpv3996         Hosp     2017-01-29                insulin 2017-01-29
#> 140 jexpv3996         Hosp     2017-01-29                   ARBs 2017-01-29
#> 141 jexpv3996         Hosp     2017-01-29         Anticoagulants 2017-01-29
#> 142 jexpv3996         Hosp     2017-01-29             Biguanides 2017-01-29
#> 143 jexpv3996         Hosp     2017-01-29 antiplatelet medicines 2017-01-29
#> 144 jexpv3996         Hosp     2017-01-29              Glinifera 2017-01-29
#> 145 jexpv3996         Hosp     2017-01-29          Sulfonylureas 2017-01-29
#> 146 jexpv3996         Hosp     2017-01-29          beta-blockers 2017-01-29
#> 147 gzqth4176         Hosp     2017-03-15                Statins 2017-03-15
#> 148 gzqth4176         Hosp     2017-03-15                   ARBs 2017-03-15
#> 149 gzqth4176         Hosp     2017-03-15          beta-blockers 2017-03-15
#> 150 gzqth4176         Hosp     2017-03-15          Sulfonylureas 2017-03-15
#> 151 ufrzf5777         Hosp     2017-09-18 Lipid-regulating drugs 2017-09-18
#> 152 ufrzf5777         Hosp     2017-09-18 antiplatelet medicines 2017-09-18
#> 153 ufrzf5777         Hosp     2017-09-18          beta-blockers 2017-09-18
#> 154 ufrzf5777         Hosp     2017-09-18                insulin 2017-09-18
#> 155 ufrzf5777         Hosp     2017-09-18         Anticoagulants 2017-09-18
#> 156 ufrzf5777         Hosp     2017-09-18              Glinifera 2017-09-18
#> 157 ufrzf5777         Hosp     2017-09-18          Sulfonylureas 2017-09-18
#> 158 ufrzf5777         Hosp     2017-09-18                   ARBs 2017-09-18
#> 159 ufrzf5777         Hosp     2017-09-18                Statins 2017-09-18
#> 160 cxpvu1806         Hosp     2018-06-13 Lipid-regulating drugs 2018-06-13
#> 161 cxpvu1806         Hosp     2018-06-13          Sulfonylureas 2018-06-13
#> 162 cxpvu1806         Hosp     2018-06-13                insulin 2018-06-13
#> 163 cxpvu1806         Hosp     2018-06-13 antiplatelet medicines 2018-06-13
#> 164 cxpvu1806         Hosp     2018-06-13                   ARBs 2018-06-13
#> 165 cxpvu1806         Hosp     2018-06-13             Biguanides 2018-06-13
#> 166 cxpvu1806         Hosp     2018-06-13              Glinifera 2018-06-13
#> 167 cxpvu1806         Hosp     2018-06-13         Anticoagulants 2018-06-13
#> 168 cxpvu1806         Hosp     2018-06-13          beta-blockers 2018-06-13
#> 169 cxpvu1806         Hosp     2018-06-13                Statins 2018-06-13
#> 170 iikmu1400         outp     2018-08-07             Biguanides 2018-08-07
#> 171 iikmu1400         outp     2018-08-07                insulin 2018-08-07
#> 172 iikmu1400         outp     2018-08-07              Glinifera 2018-08-07
#> 173 iikmu1400         outp     2018-08-07 antiplatelet medicines 2018-08-07
#> 174 iikmu1400         outp     2018-08-07         Anticoagulants 2018-08-07
#> 175 iikmu1400         outp     2018-08-07                Statins 2018-08-07
#> 176 yhalt3290         Hosp     2018-08-21          beta-blockers 2018-08-21
#> 177 yhalt3290         Hosp     2018-08-21                   ARBs 2018-08-21
#> 178 yhalt3290         Hosp     2018-08-21         Anticoagulants 2018-08-21
#> 179 yhalt3290         Hosp     2018-08-21          Sulfonylureas 2018-08-21
#> 180 yhalt3290         Hosp     2018-08-21              Glinifera 2018-08-21
#> 181 yhalt3290         Hosp     2018-08-21                insulin 2018-08-21
#> 182 yhalt3290         Hosp     2018-08-21 Lipid-regulating drugs 2018-08-21
#> 183 yhalt3290         Hosp     2018-08-21             Biguanides 2018-08-21
#> 184 yhalt3290         Hosp     2018-08-21                Statins 2018-08-21
#> 185 wpcud1752         Hosp     2018-10-02          beta-blockers 2018-10-02
#> 186 wpcud1752         Hosp     2018-10-02                   ARBs 2018-10-02
#> 187 wpcud1752         Hosp     2018-10-02                insulin 2018-10-02
#> 188 wpcud1752         Hosp     2018-10-02 antiplatelet medicines 2018-10-02
#> 189 wpcud1752         Hosp     2018-10-02          Sulfonylureas 2018-10-02
#> 190 wpcud1752         Hosp     2018-10-02             Biguanides 2018-10-02
#> 191 wpcud1752         Hosp     2018-10-02         Anticoagulants 2018-10-02
#> 192 wpcud1752         Hosp     2018-10-02              Glinifera 2018-10-02
#> 193 wpcud1752         Hosp     2018-10-02 Lipid-regulating drugs 2018-10-02
#> 194 wpcud1752         Hosp     2018-10-02                Statins 2018-10-02
#> 195 urnuv5242         Hosp     2018-10-06          beta-blockers 2018-10-06
#> 196 urnuv5242         Hosp     2018-10-06 Lipid-regulating drugs 2018-10-06
#> 197 urnuv5242         Hosp     2018-10-06                   ARBs 2018-10-06
#> 198 urnuv5242         Hosp     2018-10-06         Anticoagulants 2018-10-06
#> 199 urnuv5242         Hosp     2018-10-06                Statins 2018-10-06
#> 200 urnuv5242         Hosp     2018-10-06 antiplatelet medicines 2018-10-06
#> 201 wnvqe5313         Hosp     2018-10-16 antiplatelet medicines 2018-10-16
#> 202 wnvqe5313         Hosp     2018-10-16          Sulfonylureas 2018-10-16
#> 203 wnvqe5313         Hosp     2018-10-16 Lipid-regulating drugs 2018-10-16
#> 204 wnvqe5313         Hosp     2018-10-16             Biguanides 2018-10-16
#> 205 wnvqe5313         Hosp     2018-10-16         Anticoagulants 2018-10-16
#> 206 wnvqe5313         Hosp     2018-10-16                   ARBs 2018-10-16
#> 207 wnvqe5313         Hosp     2018-10-16                Statins 2018-10-16
#> 208 wnvqe5313         Hosp     2018-10-16                insulin 2018-10-16
#> 209 wnvqe5313         Hosp     2018-10-16              Glinifera 2018-10-16
#> 210 yqbkk1784         outp     2019-01-05              Glinifera 2019-01-05
#> 211 yqbkk1784         outp     2019-01-05          Sulfonylureas 2019-01-05
#> 212 yqbkk1784         outp     2019-01-05                Statins 2019-01-05
#> 213 yqbkk1784         outp     2019-01-05                   ARBs 2019-01-05
#> 214 yqbkk1784         outp     2019-01-05 antiplatelet medicines 2019-01-05
#> 215 yqbkk1784         outp     2019-01-05 Lipid-regulating drugs 2019-01-05
#> 216 yqbkk1784         outp     2019-01-05         Anticoagulants 2019-01-05
#> 217 yqbkk1784         outp     2019-01-05             Biguanides 2019-01-05
#> 218 yqbkk1784         outp     2019-01-05          beta-blockers 2019-01-05
#> 219 jtfzn4085         Hosp     2019-07-25         Anticoagulants 2019-07-25
#> 220 jtfzn4085         Hosp     2019-07-25             Biguanides 2019-07-25
#> 221 jtfzn4085         Hosp     2019-07-25                Statins 2019-07-25
#> 222 jtfzn4085         Hosp     2019-07-25          beta-blockers 2019-07-25
#> 223 jtfzn4085         Hosp     2019-07-25 antiplatelet medicines 2019-07-25
#> 224 sjkwd4201         Hosp     2019-08-09          Sulfonylureas 2019-08-09
#> 225 sjkwd4201         Hosp     2019-08-09                Statins 2019-08-09
#> 226 sjkwd4201         Hosp     2019-08-09              Glinifera 2019-08-09
#> 227 sjkwd4201         Hosp     2019-08-09                   ARBs 2019-08-09
#> 228 sjkwd4201         Hosp     2019-08-09         Anticoagulants 2019-08-09
#> 229 sjkwd4201         Hosp     2019-08-09             Biguanides 2019-08-09
#> 230 sjkwd4201         Hosp     2019-08-09 Lipid-regulating drugs 2019-08-09
#> 231 sjkwd4201         Hosp     2019-08-09 antiplatelet medicines 2019-08-09
#> 232 sjkwd4201         Hosp     2019-08-09          beta-blockers 2019-08-09
#> 233 sjkwd4201         Hosp     2019-08-09                insulin 2019-08-09
#> 234 pnvsp5269         Hosp     2019-08-11                Statins 2019-08-11
#> 235 pnvsp5269         Hosp     2019-08-11 Lipid-regulating drugs 2019-08-11
#> 236 pnvsp5269         Hosp     2019-08-11                   ARBs 2019-08-11
#> 237 pnvsp5269         Hosp     2019-08-11          beta-blockers 2019-08-11
#> 238 ngmep2107         outp     2020-02-05                Statins 2020-02-05
#> 239 ngmep2107         outp     2020-02-05 antiplatelet medicines 2020-02-05
#> 240 ngmep2107         outp     2020-02-05              Glinifera 2020-02-05
#> 241 ngmep2107         outp     2020-02-05          beta-blockers 2020-02-05
#> 242 ngmep2107         outp     2020-02-05         Anticoagulants 2020-02-05
#> 243 ngmep2107         outp     2020-02-05             Biguanides 2020-02-05
#> 244 ngmep2107         outp     2020-02-05 Lipid-regulating drugs 2020-02-05
#> 245 ngmep2107         outp     2020-02-05          Sulfonylureas 2020-02-05
#> 246 ngmep2107         outp     2020-02-05                   ARBs 2020-02-05
#> 247 lxlrq2954         outp     2020-02-26              Glinifera 2020-02-26
#> 248 lxlrq2954         outp     2020-02-26                insulin 2020-02-26
#> 249 lxlrq2954         outp     2020-02-26                Statins 2020-02-26
#> 250 lxlrq2954         outp     2020-02-26         Anticoagulants 2020-02-26
#> 251 lxlrq2954         outp     2020-02-26                   ARBs 2020-02-26
#> 252 lxlrq2954         outp     2020-02-26          Sulfonylureas 2020-02-26
#> 253 tjhep3934         outp     2020-05-08         Anticoagulants 2020-05-08
#> 254 tjhep3934         outp     2020-05-08             Biguanides 2020-05-08
#> 255 tjhep3934         outp     2020-05-08 Lipid-regulating drugs 2020-05-08
#> 256 tjhep3934         outp     2020-05-08 antiplatelet medicines 2020-05-08
#> 257 tjhep3934         outp     2020-05-08          Sulfonylureas 2020-05-08
#> 258 dxagx3662         outp     2020-06-24         Anticoagulants 2020-06-24
#> 259 dxagx3662         outp     2020-06-24             Biguanides 2020-06-24
#> 260 dxagx3662         outp     2020-06-24 Lipid-regulating drugs 2020-06-24
#> 261 dxagx3662         outp     2020-06-24                Statins 2020-06-24
#> 262 dxagx3662         outp     2020-06-24          Sulfonylureas 2020-06-24
#> 263 dxagx3662         outp     2020-06-24              Glinifera 2020-06-24
#> 264 dxagx3662         outp     2020-06-24                   ARBs 2020-06-24
#> 265 dxagx3662         outp     2020-06-24                insulin 2020-06-24
#> 266 dxagx3662         outp     2020-06-24          beta-blockers 2020-06-24
#> 267 svzdx2060         outp     2020-07-27         Anticoagulants 2020-07-27
#> 268 svzdx2060         outp     2020-07-27 antiplatelet medicines 2020-07-27
#> 269 svzdx2060         outp     2020-07-27              Glinifera 2020-07-27
#> 270 svzdx2060         outp     2020-07-27          Sulfonylureas 2020-07-27
#> 271 svzdx2060         outp     2020-07-27          beta-blockers 2020-07-27
#> 272 svzdx2060         outp     2020-07-27                Statins 2020-07-27
#> 273 svzdx2060         outp     2020-07-27 Lipid-regulating drugs 2020-07-27
#> 274 svzdx2060         outp     2020-07-27                insulin 2020-07-27
#> 275 svzdx2060         outp     2020-07-27             Biguanides 2020-07-27
#> 276 svzdx2060         outp     2020-07-27                   ARBs 2020-07-27
#> 277 nitlg1679         Hosp     2020-08-17          beta-blockers 2020-08-17
#> 278 nitlg1679         Hosp     2020-08-17 antiplatelet medicines 2020-08-17
#> 279 nitlg1679         Hosp     2020-08-17 Lipid-regulating drugs 2020-08-17
#> 
#> $lab
#>      visit_id patient_type admission_date lab_name lab_va  test_date
#> 1   gxioj2964         Hosp     2014-11-20      HDL      3 2014-11-20
#> 2   gxioj2964         Hosp     2014-11-20    HbA1c      9 2014-11-20
#> 3   gxioj2964         Hosp     2014-11-20       TC      5 2014-11-20
#> 4   gxioj2964         Hosp     2014-11-20      LDL      4 2014-11-20
#> 5   gxioj2964         Hosp     2014-11-20       TG     10 2014-11-20
#> 6   dbjmj5676         Hosp     2014-12-09      LDL      3 2014-12-09
#> 7   dbjmj5676         Hosp     2014-12-09      HDL      3 2014-12-09
#> 8   dbjmj5676         Hosp     2014-12-09       TG     11 2014-12-09
#> 9   dbjmj5676         Hosp     2014-12-09      WBC     12 2014-12-09
#> 10  dbjmj5676         Hosp     2014-12-09    HbA1c      6 2014-12-09
#> 11  cqaix5066         Hosp     2014-12-23      HDL      2 2014-12-23
#> 12  cqaix5066         Hosp     2014-12-23    HbA1c     10 2014-12-23
#> 13  cqaix5066         Hosp     2014-12-23      WBC     10 2014-12-23
#> 14  ddzxl1821         outp     2014-12-24      WBC      5 2014-12-24
#> 15  ddzxl1821         outp     2014-12-24      HDL      3 2014-12-24
#> 16  ddzxl1821         outp     2014-12-24    HbA1c      3 2014-12-24
#> 17  wkmzj1037         outp     2015-01-13    HbA1c      7 2015-01-13
#> 18  wkmzj1037         outp     2015-01-13      LDL      4 2015-01-13
#> 19  wkmzj1037         outp     2015-01-13      HDL      5 2015-01-13
#> 20  wkmzj1037         outp     2015-01-13       TC      2 2015-01-13
#> 21  wkmzj1037         outp     2015-01-13       TG      9 2015-01-13
#> 22  wkmzj1037         outp     2015-01-13      WBC      8 2015-01-13
#> 23  wqudj3630         outp     2015-01-16       TG      6 2015-01-16
#> 24  wqudj3630         outp     2015-01-16    HbA1c     10 2015-01-16
#> 25  wqudj3630         outp     2015-01-16       TC      1 2015-01-16
#> 26  rugrj5988         outp     2015-02-06      WBC      7 2015-02-06
#> 27  rugrj5988         outp     2015-02-06    HbA1c     11 2015-02-06
#> 28  rugrj5988         outp     2015-02-06      HDL      1 2015-02-06
#> 29  rugrj5988         outp     2015-02-06      LDL      1 2015-02-06
#> 30  rugrj5988         outp     2015-02-06       TC      2 2015-02-06
#> 31  rugrj5988         outp     2015-02-06       TG      9 2015-02-06
#> 32  oezbz4107         Hosp     2015-02-15    HbA1c      8 2015-02-15
#> 33  oezbz4107         Hosp     2015-02-15      WBC     17 2015-02-15
#> 34  oezbz4107         Hosp     2015-02-15      HDL      1 2015-02-15
#> 35  oezbz4107         Hosp     2015-02-15      LDL      2 2015-02-15
#> 36  oezbz4107         Hosp     2015-02-15       TG      5 2015-02-15
#> 37  oezbz4107         Hosp     2015-02-15       TC      5 2015-02-15
#> 38  keksr5885         outp     2015-04-16       TC      5 2015-04-16
#> 39  keksr5885         outp     2015-04-16      WBC      5 2015-04-16
#> 40  keksr5885         outp     2015-04-16       TG      1 2015-04-16
#> 41  keksr5885         outp     2015-04-16      HDL      1 2015-04-16
#> 42  keksr5885         outp     2015-04-16      LDL      1 2015-04-16
#> 43  keksr5885         outp     2015-04-16    HbA1c     12 2015-04-16
#> 44  auphy4493         outp     2015-05-01       TC      3 2015-05-01
#> 45  auphy4493         outp     2015-05-01    HbA1c      6 2015-05-01
#> 46  auphy4493         outp     2015-05-01      WBC     19 2015-05-01
#> 47  auphy4493         outp     2015-05-01      HDL      4 2015-05-01
#> 48  auphy4493         outp     2015-05-01      LDL      1 2015-05-01
#> 49  auphy4493         outp     2015-05-01       TG     15 2015-05-01
#> 50  zfmwv1799         outp     2015-05-19       TG      6 2015-05-19
#> 51  zfmwv1799         outp     2015-05-19      LDL      5 2015-05-19
#> 52  zfmwv1799         outp     2015-05-19      HDL      2 2015-05-19
#> 53  zfmwv1799         outp     2015-05-19    HbA1c      4 2015-05-19
#> 54  zfmwv1799         outp     2015-05-19       TC      5 2015-05-19
#> 55  zfmwv1799         outp     2015-05-19      WBC     12 2015-05-19
#> 56  jqogc2293         Hosp     2015-05-23      HDL      2 2015-05-23
#> 57  jqogc2293         Hosp     2015-05-23      LDL      3 2015-05-23
#> 58  jqogc2293         Hosp     2015-05-23    HbA1c     10 2015-05-23
#> 59  cgzet3758         outp     2015-06-21    HbA1c     10 2015-06-21
#> 60  cgzet3758         outp     2015-06-21      WBC      1 2015-06-21
#> 61  cgzet3758         outp     2015-06-21      HDL      3 2015-06-21
#> 62  cgzet3758         outp     2015-06-21      LDL      5 2015-06-21
#> 63  nwdni4370         outp     2015-08-18      HDL      3 2015-08-18
#> 64  nwdni4370         outp     2015-08-18    HbA1c     10 2015-08-18
#> 65  nwdni4370         outp     2015-08-18      LDL      3 2015-08-18
#> 66  inlmf5251         Hosp     2015-11-03       TC      6 2015-11-03
#> 67  inlmf5251         Hosp     2015-11-03    HbA1c      1 2015-11-03
#> 68  inlmf5251         Hosp     2015-11-03       TG     10 2015-11-03
#> 69  inlmf5251         Hosp     2015-11-03      LDL      3 2015-11-03
#> 70  kykdz3468         Hosp     2016-01-01       TG     10 2016-01-01
#> 71  kykdz3468         Hosp     2016-01-01       TC      7 2016-01-01
#> 72  kykdz3468         Hosp     2016-01-01      WBC     19 2016-01-01
#> 73  pjdin1619         Hosp     2016-04-24       TC      1 2016-04-24
#> 74  pjdin1619         Hosp     2016-04-24    HbA1c      5 2016-04-24
#> 75  pjdin1619         Hosp     2016-04-24      WBC      3 2016-04-24
#> 76  pjdin1619         Hosp     2016-04-24      LDL      4 2016-04-24
#> 77  pjdin1619         Hosp     2016-04-24       TG      6 2016-04-24
#> 78  pjdin1619         Hosp     2016-04-24      HDL      1 2016-04-24
#> 79  dbxdg2557         outp     2016-08-23      HDL      3 2016-08-23
#> 80  dbxdg2557         outp     2016-08-23       TC      1 2016-08-23
#> 81  dbxdg2557         outp     2016-08-23      LDL      3 2016-08-23
#> 82  dbxdg2557         outp     2016-08-23    HbA1c      3 2016-08-23
#> 83  dbxdg2557         outp     2016-08-23      WBC     19 2016-08-23
#> 84  dbxdg2557         outp     2016-08-23       TG      2 2016-08-23
#> 85  mdpnk4183         outp     2016-09-08      WBC      8 2016-09-08
#> 86  mdpnk4183         outp     2016-09-08       TC      2 2016-09-08
#> 87  mdpnk4183         outp     2016-09-08      LDL      4 2016-09-08
#> 88  mdpnk4183         outp     2016-09-08       TG      4 2016-09-08
#> 89  mdpnk4183         outp     2016-09-08      HDL      4 2016-09-08
#> 90  mdpnk4183         outp     2016-09-08    HbA1c      3 2016-09-08
#> 91  wzcdr3096         outp     2017-01-04      HDL      3 2017-01-04
#> 92  wzcdr3096         outp     2017-01-04       TC      1 2017-01-04
#> 93  wzcdr3096         outp     2017-01-04      WBC     17 2017-01-04
#> 94  wzcdr3096         outp     2017-01-04       TG     11 2017-01-04
#> 95  wzcdr3096         outp     2017-01-04      LDL      5 2017-01-04
#> 96  wzcdr3096         outp     2017-01-04    HbA1c      1 2017-01-04
#> 97  jexpv3996         Hosp     2017-01-29      HDL      5 2017-01-29
#> 98  jexpv3996         Hosp     2017-01-29      LDL      3 2017-01-29
#> 99  jexpv3996         Hosp     2017-01-29       TG      5 2017-01-29
#> 100 gzqth4176         Hosp     2017-03-15      HDL      3 2017-03-15
#> 101 gzqth4176         Hosp     2017-03-15      WBC      2 2017-03-15
#> 102 gzqth4176         Hosp     2017-03-15       TG     10 2017-03-15
#> 103 gzqth4176         Hosp     2017-03-15      LDL      1 2017-03-15
#> 104 gzqth4176         Hosp     2017-03-15       TC      4 2017-03-15
#> 105 ufrzf5777         Hosp     2017-09-18       TC      7 2017-09-18
#> 106 ufrzf5777         Hosp     2017-09-18      WBC      8 2017-09-18
#> 107 ufrzf5777         Hosp     2017-09-18    HbA1c      3 2017-09-18
#> 108 ufrzf5777         Hosp     2017-09-18      LDL      5 2017-09-18
#> 109 cxpvu1806         Hosp     2018-06-13      HDL      3 2018-06-13
#> 110 cxpvu1806         Hosp     2018-06-13      LDL      3 2018-06-13
#> 111 cxpvu1806         Hosp     2018-06-13       TG      8 2018-06-13
#> 112 cxpvu1806         Hosp     2018-06-13    HbA1c      5 2018-06-13
#> 113 iikmu1400         outp     2018-08-07      WBC      1 2018-08-07
#> 114 iikmu1400         outp     2018-08-07    HbA1c     10 2018-08-07
#> 115 iikmu1400         outp     2018-08-07      LDL      5 2018-08-07
#> 116 iikmu1400         outp     2018-08-07       TC      3 2018-08-07
#> 117 iikmu1400         outp     2018-08-07      HDL      2 2018-08-07
#> 118 iikmu1400         outp     2018-08-07       TG     18 2018-08-07
#> 119 yhalt3290         Hosp     2018-08-21       TC      4 2018-08-21
#> 120 yhalt3290         Hosp     2018-08-21    HbA1c     12 2018-08-21
#> 121 yhalt3290         Hosp     2018-08-21       TG      9 2018-08-21
#> 122 wpcud1752         Hosp     2018-10-02    HbA1c      6 2018-10-02
#> 123 wpcud1752         Hosp     2018-10-02      WBC     17 2018-10-02
#> 124 wpcud1752         Hosp     2018-10-02      LDL      2 2018-10-02
#> 125 urnuv5242         Hosp     2018-10-06      WBC      1 2018-10-06
#> 126 urnuv5242         Hosp     2018-10-06       TC      2 2018-10-06
#> 127 urnuv5242         Hosp     2018-10-06      HDL      2 2018-10-06
#> 128 urnuv5242         Hosp     2018-10-06       TG      7 2018-10-06
#> 129 urnuv5242         Hosp     2018-10-06    HbA1c      5 2018-10-06
#> 130 wnvqe5313         Hosp     2018-10-16      HDL      2 2018-10-16
#> 131 wnvqe5313         Hosp     2018-10-16       TC      2 2018-10-16
#> 132 wnvqe5313         Hosp     2018-10-16      LDL      5 2018-10-16
#> 133 wnvqe5313         Hosp     2018-10-16    HbA1c      7 2018-10-16
#> 134 yqbkk1784         outp     2019-01-05       TC      7 2019-01-05
#> 135 yqbkk1784         outp     2019-01-05       TG     20 2019-01-05
#> 136 yqbkk1784         outp     2019-01-05      WBC      5 2019-01-05
#> 137 yqbkk1784         outp     2019-01-05    HbA1c     11 2019-01-05
#> 138 yqbkk1784         outp     2019-01-05      HDL      2 2019-01-05
#> 139 jtfzn4085         Hosp     2019-07-25    HbA1c      4 2019-07-25
#> 140 jtfzn4085         Hosp     2019-07-25       TC      2 2019-07-25
#> 141 jtfzn4085         Hosp     2019-07-25      LDL      4 2019-07-25
#> 142 jtfzn4085         Hosp     2019-07-25       TG      4 2019-07-25
#> 143 sjkwd4201         Hosp     2019-08-09       TG      3 2019-08-09
#> 144 sjkwd4201         Hosp     2019-08-09    HbA1c      3 2019-08-09
#> 145 sjkwd4201         Hosp     2019-08-09      HDL      3 2019-08-09
#> 146 sjkwd4201         Hosp     2019-08-09      LDL      5 2019-08-09
#> 147 pnvsp5269         Hosp     2019-08-11      WBC     17 2019-08-11
#> 148 pnvsp5269         Hosp     2019-08-11       TC      2 2019-08-11
#> 149 pnvsp5269         Hosp     2019-08-11      HDL      3 2019-08-11
#> 150 pnvsp5269         Hosp     2019-08-11      LDL      3 2019-08-11
#> 151 ngmep2107         outp     2020-02-05    HbA1c      3 2020-02-05
#> 152 ngmep2107         outp     2020-02-05      WBC     17 2020-02-05
#> 153 ngmep2107         outp     2020-02-05       TC      3 2020-02-05
#> 154 ngmep2107         outp     2020-02-05      HDL      4 2020-02-05
#> 155 lxlrq2954         outp     2020-02-26      HDL      1 2020-02-26
#> 156 lxlrq2954         outp     2020-02-26    HbA1c      8 2020-02-26
#> 157 lxlrq2954         outp     2020-02-26      WBC     11 2020-02-26
#> 158 lxlrq2954         outp     2020-02-26       TC      5 2020-02-26
#> 159 tjhep3934         outp     2020-05-08       TG     15 2020-05-08
#> 160 tjhep3934         outp     2020-05-08    HbA1c     11 2020-05-08
#> 161 tjhep3934         outp     2020-05-08      HDL      3 2020-05-08
#> 162 tjhep3934         outp     2020-05-08      LDL      4 2020-05-08
#> 163 tjhep3934         outp     2020-05-08       TC      2 2020-05-08
#> 164 dxagx3662         outp     2020-06-24      HDL      2 2020-06-24
#> 165 dxagx3662         outp     2020-06-24    HbA1c      8 2020-06-24
#> 166 dxagx3662         outp     2020-06-24       TC      6 2020-06-24
#> 167 dxagx3662         outp     2020-06-24       TG      7 2020-06-24
#> 168 dxagx3662         outp     2020-06-24      WBC     13 2020-06-24
#> 169 dxagx3662         outp     2020-06-24      LDL      3 2020-06-24
#> 170 svzdx2060         outp     2020-07-27       TG     10 2020-07-27
#> 171 svzdx2060         outp     2020-07-27      WBC     13 2020-07-27
#> 172 svzdx2060         outp     2020-07-27      HDL      3 2020-07-27
#> 173 svzdx2060         outp     2020-07-27      LDL      2 2020-07-27
#> 174 nitlg1679         Hosp     2020-08-17       TG     13 2020-08-17
#> 175 nitlg1679         Hosp     2020-08-17      LDL      4 2020-08-17
#> 176 nitlg1679         Hosp     2020-08-17    HbA1c     12 2020-08-17
#> 177 nitlg1679         Hosp     2020-08-17      WBC     14 2020-08-17
#> 178 nitlg1679         Hosp     2020-08-17      HDL      5 2020-08-17
#> 
#> $diag
#> # A tibble: 39 × 5
#>    visit_id  patient_type admission_date diag_date  diag_union                  
#>    <chr>     <chr>        <date>         <date>     <chr>                       
#>  1 gxioj2964 Hosp         2014-11-20     2014-11-20 Symptomatic neuropathy,Stro…
#>  2 dbjmj5676 Hosp         2014-12-09     2014-12-09 Dementia,Connective tissue …
#>  3 cqaix5066 Hosp         2014-12-23     2014-12-23 Mild liver disease,Dementia…
#>  4 ddzxl1821 outp         2014-12-24     2014-12-24 NA                          
#>  5 wkmzj1037 outp         2015-01-13     2015-01-13 NA,Dementia,Moderate and se…
#>  6 wqudj3630 outp         2015-01-16     2015-01-16 Lymphatic carcinoma,Moderat…
#>  7 rugrj5988 outp         2015-02-06     2015-02-06 NA,diabetes,myocardial infa…
#>  8 oezbz4107 Hosp         2015-02-15     2015-02-15 diabetes,Mild liver disease…
#>  9 keksr5885 outp         2015-04-16     2015-04-16 Lymphatic carcinoma,NA,Mild…
#> 10 auphy4493 outp         2015-05-01     2015-05-01 high blood pressure,Mild li…
#> # … with 29 more rows
#>