calculate the virtual dataset of a patient

calculate the virtual dataset of a patient

Usage

Virtual_med_one(
  birth_start_date,
  birth_end_date,
  study_start_date,
  study_end_date,
  average_year_visit_n
)

Arguments

birth_start_date

Minimum expected birth date of the patient.

birth_end_date

Maximum expected birth date of the patient.

study_start_date

The starting time of the baseline period specified in the study.

study_end_date

The end time of the follow-up period specified in the study

average_year_visit_n

The average number of visits by patients each year is 6 by default. The data are from the statistical records of the Chinese people.

Value

Returns a data set of a patients, including a doctor's table, prescription table, diagnosis table, and biochemical test table.

Examples

Virtual_med_one(
 birth_start_date = "1980-01-01",birth_end_date ="2000-12-31" ,
   study_start_date = "2014-01-01",study_end_date ="2020-12-31",
   average_year_visit_n = 6)
#> $pat
#>     visit_id    sex  birthdate patient_type admission_date Discharge_date
#> 1  fnhjl5940 female 1987-05-01         Hosp     2014-01-22     2014-02-28
#> 2  ieetc5356 female 1987-05-01         Hosp     2014-02-14     2014-03-23
#> 3  svpiu3306 female 1987-05-01         Hosp     2014-07-07     2014-08-13
#> 4  ytyga3981 female 1987-05-01         outp     2014-10-18     2014-11-24
#> 5  tntmb2887 female 1987-05-01         Hosp     2014-11-27     2015-01-03
#> 6  qnoqv3107 female 1987-05-01         outp     2015-01-25     2015-03-03
#> 7  fmsbq3634 female 1987-05-01         Hosp     2015-01-26     2015-03-04
#> 8  fheez3668 female 1987-05-01         Hosp     2015-05-20     2015-06-26
#> 9  qrcql3396 female 1987-05-01         outp     2015-06-27     2015-08-03
#> 10 illqe2480 female 1987-05-01         Hosp     2016-01-26     2016-03-03
#> 11 wczwb4690 female 1987-05-01         outp     2016-03-08     2016-04-14
#> 12 hfopx3926 female 1987-05-01         Hosp     2016-03-27     2016-05-03
#> 13 klzyf4532 female 1987-05-01         Hosp     2016-05-31     2016-07-07
#> 14 ojava5964 female 1987-05-01         Hosp     2016-08-13     2016-09-19
#> 15 skmtg5971 female 1987-05-01         outp     2017-01-17     2017-02-23
#> 16 beklf3627 female 1987-05-01         Hosp     2017-06-29     2017-08-05
#> 17 bvlns4639 female 1987-05-01         outp     2017-07-02     2017-08-08
#> 18 dallj4117 female 1987-05-01         Hosp     2017-07-21     2017-08-27
#> 19 egzws5800 female 1987-05-01         Hosp     2017-09-02     2017-10-09
#> 20 wnlkn3812 female 1987-05-01         outp     2017-10-31     2017-12-07
#> 21 hkwat1476 female 1987-05-01         Hosp     2018-04-30     2018-06-06
#> 22 vrpfw3655 female 1987-05-01         Hosp     2018-05-07     2018-06-13
#> 23 fbdhh2630 female 1987-05-01         outp     2018-06-03     2018-07-10
#> 24 ilgvm4712 female 1987-05-01         Hosp     2018-09-30     2018-11-06
#> 25 dfdtu5442 female 1987-05-01         outp     2019-04-20     2019-05-27
#> 26 bfwzd1905 female 1987-05-01         outp     2019-05-23     2019-06-29
#> 27 pwsks4093 female 1987-05-01         outp     2019-06-02     2019-07-09
#> 28 pjeto4451 female 1987-05-01         outp     2019-07-20     2019-08-26
#> 29 zsfbm3006 female 1987-05-01         outp     2019-10-18     2019-11-24
#> 30 gsutd5982 female 1987-05-01         outp     2019-12-02     2020-01-08
#> 31 qmwji3541 female 1987-05-01         outp     2019-12-08     2020-01-14
#> 32 smben1421 female 1987-05-01         Hosp     2020-01-07     2020-02-13
#> 33 kuazo2021 female 1987-05-01         outp     2020-01-16     2020-02-22
#> 34 pmcrk4713 female 1987-05-01         Hosp     2020-03-24     2020-04-30
#> 35 xpixr3135 female 1987-05-01         Hosp     2020-04-26     2020-06-02
#> 36 twdrf4539 female 1987-05-01         outp     2020-05-18     2020-06-24
#> 37 svjhc1783 female 1987-05-01         outp     2020-06-04     2020-07-11
#> 38 djiue5325 female 1987-05-01         outp     2020-06-25     2020-08-01
#> 39 cknsq2879 female 1987-05-01         Hosp     2020-07-18     2020-08-24
#> 40 tkpft4981 female 1987-05-01         Hosp     2020-10-25     2020-12-01
#> 
#> $prec
#>      visit_id patient_type admission_date              prec_name  prec_date
#> 1   fnhjl5940         Hosp     2014-01-22          beta-blockers 2014-01-22
#> 2   fnhjl5940         Hosp     2014-01-22             Biguanides 2014-01-22
#> 3   fnhjl5940         Hosp     2014-01-22                   ARBs 2014-01-22
#> 4   fnhjl5940         Hosp     2014-01-22                Statins 2014-01-22
#> 5   fnhjl5940         Hosp     2014-01-22              Glinifera 2014-01-22
#> 6   fnhjl5940         Hosp     2014-01-22 antiplatelet medicines 2014-01-22
#> 7   ieetc5356         Hosp     2014-02-14                Statins 2014-02-14
#> 8   ieetc5356         Hosp     2014-02-14                   ARBs 2014-02-14
#> 9   ieetc5356         Hosp     2014-02-14         Anticoagulants 2014-02-14
#> 10  ieetc5356         Hosp     2014-02-14          beta-blockers 2014-02-14
#> 11  ieetc5356         Hosp     2014-02-14              Glinifera 2014-02-14
#> 12  ieetc5356         Hosp     2014-02-14 Lipid-regulating drugs 2014-02-14
#> 13  ieetc5356         Hosp     2014-02-14             Biguanides 2014-02-14
#> 14  svpiu3306         Hosp     2014-07-07              Glinifera 2014-07-07
#> 15  svpiu3306         Hosp     2014-07-07         Anticoagulants 2014-07-07
#> 16  svpiu3306         Hosp     2014-07-07          Sulfonylureas 2014-07-07
#> 17  svpiu3306         Hosp     2014-07-07 Lipid-regulating drugs 2014-07-07
#> 18  ytyga3981         outp     2014-10-18 Lipid-regulating drugs 2014-10-18
#> 19  ytyga3981         outp     2014-10-18 antiplatelet medicines 2014-10-18
#> 20  ytyga3981         outp     2014-10-18                Statins 2014-10-18
#> 21  ytyga3981         outp     2014-10-18             Biguanides 2014-10-18
#> 22  ytyga3981         outp     2014-10-18         Anticoagulants 2014-10-18
#> 23  tntmb2887         Hosp     2014-11-27 antiplatelet medicines 2014-11-27
#> 24  tntmb2887         Hosp     2014-11-27                Statins 2014-11-27
#> 25  tntmb2887         Hosp     2014-11-27              Glinifera 2014-11-27
#> 26  tntmb2887         Hosp     2014-11-27 Lipid-regulating drugs 2014-11-27
#> 27  qnoqv3107         outp     2015-01-25              Glinifera 2015-01-25
#> 28  qnoqv3107         outp     2015-01-25 Lipid-regulating drugs 2015-01-25
#> 29  qnoqv3107         outp     2015-01-25                   ARBs 2015-01-25
#> 30  qnoqv3107         outp     2015-01-25          beta-blockers 2015-01-25
#> 31  fmsbq3634         Hosp     2015-01-26                   ARBs 2015-01-26
#> 32  fmsbq3634         Hosp     2015-01-26                Statins 2015-01-26
#> 33  fmsbq3634         Hosp     2015-01-26         Anticoagulants 2015-01-26
#> 34  fheez3668         Hosp     2015-05-20         Anticoagulants 2015-05-20
#> 35  fheez3668         Hosp     2015-05-20              Glinifera 2015-05-20
#> 36  fheez3668         Hosp     2015-05-20                insulin 2015-05-20
#> 37  fheez3668         Hosp     2015-05-20                Statins 2015-05-20
#> 38  fheez3668         Hosp     2015-05-20             Biguanides 2015-05-20
#> 39  fheez3668         Hosp     2015-05-20          beta-blockers 2015-05-20
#> 40  fheez3668         Hosp     2015-05-20 Lipid-regulating drugs 2015-05-20
#> 41  fheez3668         Hosp     2015-05-20                   ARBs 2015-05-20
#> 42  fheez3668         Hosp     2015-05-20          Sulfonylureas 2015-05-20
#> 43  qrcql3396         outp     2015-06-27                   ARBs 2015-06-27
#> 44  qrcql3396         outp     2015-06-27             Biguanides 2015-06-27
#> 45  qrcql3396         outp     2015-06-27 Lipid-regulating drugs 2015-06-27
#> 46  qrcql3396         outp     2015-06-27              Glinifera 2015-06-27
#> 47  qrcql3396         outp     2015-06-27                Statins 2015-06-27
#> 48  qrcql3396         outp     2015-06-27                insulin 2015-06-27
#> 49  illqe2480         Hosp     2016-01-26                   ARBs 2016-01-26
#> 50  illqe2480         Hosp     2016-01-26 antiplatelet medicines 2016-01-26
#> 51  illqe2480         Hosp     2016-01-26          beta-blockers 2016-01-26
#> 52  illqe2480         Hosp     2016-01-26                insulin 2016-01-26
#> 53  illqe2480         Hosp     2016-01-26             Biguanides 2016-01-26
#> 54  wczwb4690         outp     2016-03-08                insulin 2016-03-08
#> 55  wczwb4690         outp     2016-03-08              Glinifera 2016-03-08
#> 56  wczwb4690         outp     2016-03-08             Biguanides 2016-03-08
#> 57  wczwb4690         outp     2016-03-08         Anticoagulants 2016-03-08
#> 58  wczwb4690         outp     2016-03-08 antiplatelet medicines 2016-03-08
#> 59  hfopx3926         Hosp     2016-03-27             Biguanides 2016-03-27
#> 60  hfopx3926         Hosp     2016-03-27                   ARBs 2016-03-27
#> 61  hfopx3926         Hosp     2016-03-27              Glinifera 2016-03-27
#> 62  hfopx3926         Hosp     2016-03-27          Sulfonylureas 2016-03-27
#> 63  hfopx3926         Hosp     2016-03-27                Statins 2016-03-27
#> 64  hfopx3926         Hosp     2016-03-27                insulin 2016-03-27
#> 65  hfopx3926         Hosp     2016-03-27 Lipid-regulating drugs 2016-03-27
#> 66  hfopx3926         Hosp     2016-03-27 antiplatelet medicines 2016-03-27
#> 67  hfopx3926         Hosp     2016-03-27          beta-blockers 2016-03-27
#> 68  hfopx3926         Hosp     2016-03-27         Anticoagulants 2016-03-27
#> 69  klzyf4532         Hosp     2016-05-31 antiplatelet medicines 2016-05-31
#> 70  klzyf4532         Hosp     2016-05-31         Anticoagulants 2016-05-31
#> 71  klzyf4532         Hosp     2016-05-31 Lipid-regulating drugs 2016-05-31
#> 72  klzyf4532         Hosp     2016-05-31                insulin 2016-05-31
#> 73  ojava5964         Hosp     2016-08-13          Sulfonylureas 2016-08-13
#> 74  ojava5964         Hosp     2016-08-13                Statins 2016-08-13
#> 75  ojava5964         Hosp     2016-08-13          beta-blockers 2016-08-13
#> 76  ojava5964         Hosp     2016-08-13                   ARBs 2016-08-13
#> 77  ojava5964         Hosp     2016-08-13 antiplatelet medicines 2016-08-13
#> 78  ojava5964         Hosp     2016-08-13         Anticoagulants 2016-08-13
#> 79  ojava5964         Hosp     2016-08-13             Biguanides 2016-08-13
#> 80  ojava5964         Hosp     2016-08-13                insulin 2016-08-13
#> 81  skmtg5971         outp     2017-01-17 antiplatelet medicines 2017-01-17
#> 82  skmtg5971         outp     2017-01-17          Sulfonylureas 2017-01-17
#> 83  skmtg5971         outp     2017-01-17          beta-blockers 2017-01-17
#> 84  skmtg5971         outp     2017-01-17                Statins 2017-01-17
#> 85  skmtg5971         outp     2017-01-17              Glinifera 2017-01-17
#> 86  skmtg5971         outp     2017-01-17 Lipid-regulating drugs 2017-01-17
#> 87  beklf3627         Hosp     2017-06-29          beta-blockers 2017-06-29
#> 88  beklf3627         Hosp     2017-06-29              Glinifera 2017-06-29
#> 89  beklf3627         Hosp     2017-06-29 antiplatelet medicines 2017-06-29
#> 90  beklf3627         Hosp     2017-06-29                insulin 2017-06-29
#> 91  beklf3627         Hosp     2017-06-29                Statins 2017-06-29
#> 92  beklf3627         Hosp     2017-06-29          Sulfonylureas 2017-06-29
#> 93  beklf3627         Hosp     2017-06-29             Biguanides 2017-06-29
#> 94  beklf3627         Hosp     2017-06-29                   ARBs 2017-06-29
#> 95  beklf3627         Hosp     2017-06-29         Anticoagulants 2017-06-29
#> 96  bvlns4639         outp     2017-07-02                insulin 2017-07-02
#> 97  bvlns4639         outp     2017-07-02         Anticoagulants 2017-07-02
#> 98  bvlns4639         outp     2017-07-02 Lipid-regulating drugs 2017-07-02
#> 99  bvlns4639         outp     2017-07-02             Biguanides 2017-07-02
#> 100 dallj4117         Hosp     2017-07-21              Glinifera 2017-07-21
#> 101 dallj4117         Hosp     2017-07-21                   ARBs 2017-07-21
#> 102 dallj4117         Hosp     2017-07-21         Anticoagulants 2017-07-21
#> 103 dallj4117         Hosp     2017-07-21 antiplatelet medicines 2017-07-21
#> 104 dallj4117         Hosp     2017-07-21                Statins 2017-07-21
#> 105 dallj4117         Hosp     2017-07-21          Sulfonylureas 2017-07-21
#> 106 egzws5800         Hosp     2017-09-02          Sulfonylureas 2017-09-02
#> 107 egzws5800         Hosp     2017-09-02              Glinifera 2017-09-02
#> 108 egzws5800         Hosp     2017-09-02                   ARBs 2017-09-02
#> 109 egzws5800         Hosp     2017-09-02          beta-blockers 2017-09-02
#> 110 wnlkn3812         outp     2017-10-31         Anticoagulants 2017-10-31
#> 111 wnlkn3812         outp     2017-10-31                insulin 2017-10-31
#> 112 wnlkn3812         outp     2017-10-31          beta-blockers 2017-10-31
#> 113 hkwat1476         Hosp     2018-04-30             Biguanides 2018-04-30
#> 114 hkwat1476         Hosp     2018-04-30          Sulfonylureas 2018-04-30
#> 115 hkwat1476         Hosp     2018-04-30                insulin 2018-04-30
#> 116 hkwat1476         Hosp     2018-04-30          beta-blockers 2018-04-30
#> 117 hkwat1476         Hosp     2018-04-30 Lipid-regulating drugs 2018-04-30
#> 118 hkwat1476         Hosp     2018-04-30              Glinifera 2018-04-30
#> 119 hkwat1476         Hosp     2018-04-30                Statins 2018-04-30
#> 120 hkwat1476         Hosp     2018-04-30         Anticoagulants 2018-04-30
#> 121 hkwat1476         Hosp     2018-04-30                   ARBs 2018-04-30
#> 122 vrpfw3655         Hosp     2018-05-07 Lipid-regulating drugs 2018-05-07
#> 123 vrpfw3655         Hosp     2018-05-07                insulin 2018-05-07
#> 124 vrpfw3655         Hosp     2018-05-07          beta-blockers 2018-05-07
#> 125 vrpfw3655         Hosp     2018-05-07                Statins 2018-05-07
#> 126 fbdhh2630         outp     2018-06-03         Anticoagulants 2018-06-03
#> 127 fbdhh2630         outp     2018-06-03                insulin 2018-06-03
#> 128 fbdhh2630         outp     2018-06-03             Biguanides 2018-06-03
#> 129 fbdhh2630         outp     2018-06-03 antiplatelet medicines 2018-06-03
#> 130 ilgvm4712         Hosp     2018-09-30             Biguanides 2018-09-30
#> 131 ilgvm4712         Hosp     2018-09-30                Statins 2018-09-30
#> 132 ilgvm4712         Hosp     2018-09-30 antiplatelet medicines 2018-09-30
#> 133 ilgvm4712         Hosp     2018-09-30 Lipid-regulating drugs 2018-09-30
#> 134 ilgvm4712         Hosp     2018-09-30          Sulfonylureas 2018-09-30
#> 135 ilgvm4712         Hosp     2018-09-30                insulin 2018-09-30
#> 136 ilgvm4712         Hosp     2018-09-30                   ARBs 2018-09-30
#> 137 ilgvm4712         Hosp     2018-09-30         Anticoagulants 2018-09-30
#> 138 dfdtu5442         outp     2019-04-20              Glinifera 2019-04-20
#> 139 dfdtu5442         outp     2019-04-20          beta-blockers 2019-04-20
#> 140 dfdtu5442         outp     2019-04-20                Statins 2019-04-20
#> 141 dfdtu5442         outp     2019-04-20 antiplatelet medicines 2019-04-20
#> 142 dfdtu5442         outp     2019-04-20             Biguanides 2019-04-20
#> 143 dfdtu5442         outp     2019-04-20                   ARBs 2019-04-20
#> 144 dfdtu5442         outp     2019-04-20                insulin 2019-04-20
#> 145 dfdtu5442         outp     2019-04-20          Sulfonylureas 2019-04-20
#> 146 dfdtu5442         outp     2019-04-20         Anticoagulants 2019-04-20
#> 147 dfdtu5442         outp     2019-04-20 Lipid-regulating drugs 2019-04-20
#> 148 bfwzd1905         outp     2019-05-23          Sulfonylureas 2019-05-23
#> 149 bfwzd1905         outp     2019-05-23                insulin 2019-05-23
#> 150 bfwzd1905         outp     2019-05-23              Glinifera 2019-05-23
#> 151 bfwzd1905         outp     2019-05-23             Biguanides 2019-05-23
#> 152 bfwzd1905         outp     2019-05-23                   ARBs 2019-05-23
#> 153 bfwzd1905         outp     2019-05-23          beta-blockers 2019-05-23
#> 154 bfwzd1905         outp     2019-05-23         Anticoagulants 2019-05-23
#> 155 pwsks4093         outp     2019-06-02          beta-blockers 2019-06-02
#> 156 pwsks4093         outp     2019-06-02 Lipid-regulating drugs 2019-06-02
#> 157 pwsks4093         outp     2019-06-02                insulin 2019-06-02
#> 158 pwsks4093         outp     2019-06-02              Glinifera 2019-06-02
#> 159 pwsks4093         outp     2019-06-02         Anticoagulants 2019-06-02
#> 160 pwsks4093         outp     2019-06-02             Biguanides 2019-06-02
#> 161 pjeto4451         outp     2019-07-20         Anticoagulants 2019-07-20
#> 162 pjeto4451         outp     2019-07-20                   ARBs 2019-07-20
#> 163 pjeto4451         outp     2019-07-20 Lipid-regulating drugs 2019-07-20
#> 164 pjeto4451         outp     2019-07-20                Statins 2019-07-20
#> 165 pjeto4451         outp     2019-07-20          Sulfonylureas 2019-07-20
#> 166 pjeto4451         outp     2019-07-20          beta-blockers 2019-07-20
#> 167 pjeto4451         outp     2019-07-20                insulin 2019-07-20
#> 168 pjeto4451         outp     2019-07-20 antiplatelet medicines 2019-07-20
#> 169 pjeto4451         outp     2019-07-20              Glinifera 2019-07-20
#> 170 pjeto4451         outp     2019-07-20             Biguanides 2019-07-20
#> 171 zsfbm3006         outp     2019-10-18 Lipid-regulating drugs 2019-10-18
#> 172 zsfbm3006         outp     2019-10-18              Glinifera 2019-10-18
#> 173 zsfbm3006         outp     2019-10-18         Anticoagulants 2019-10-18
#> 174 zsfbm3006         outp     2019-10-18          beta-blockers 2019-10-18
#> 175 zsfbm3006         outp     2019-10-18             Biguanides 2019-10-18
#> 176 zsfbm3006         outp     2019-10-18                   ARBs 2019-10-18
#> 177 zsfbm3006         outp     2019-10-18                insulin 2019-10-18
#> 178 zsfbm3006         outp     2019-10-18 antiplatelet medicines 2019-10-18
#> 179 zsfbm3006         outp     2019-10-18                Statins 2019-10-18
#> 180 gsutd5982         outp     2019-12-02          beta-blockers 2019-12-02
#> 181 gsutd5982         outp     2019-12-02             Biguanides 2019-12-02
#> 182 gsutd5982         outp     2019-12-02              Glinifera 2019-12-02
#> 183 gsutd5982         outp     2019-12-02                Statins 2019-12-02
#> 184 gsutd5982         outp     2019-12-02                insulin 2019-12-02
#> 185 gsutd5982         outp     2019-12-02 Lipid-regulating drugs 2019-12-02
#> 186 qmwji3541         outp     2019-12-08         Anticoagulants 2019-12-08
#> 187 qmwji3541         outp     2019-12-08          Sulfonylureas 2019-12-08
#> 188 qmwji3541         outp     2019-12-08                insulin 2019-12-08
#> 189 qmwji3541         outp     2019-12-08                Statins 2019-12-08
#> 190 qmwji3541         outp     2019-12-08 Lipid-regulating drugs 2019-12-08
#> 191 qmwji3541         outp     2019-12-08                   ARBs 2019-12-08
#> 192 qmwji3541         outp     2019-12-08              Glinifera 2019-12-08
#> 193 qmwji3541         outp     2019-12-08          beta-blockers 2019-12-08
#> 194 qmwji3541         outp     2019-12-08             Biguanides 2019-12-08
#> 195 qmwji3541         outp     2019-12-08 antiplatelet medicines 2019-12-08
#> 196 smben1421         Hosp     2020-01-07          beta-blockers 2020-01-07
#> 197 smben1421         Hosp     2020-01-07                   ARBs 2020-01-07
#> 198 smben1421         Hosp     2020-01-07                insulin 2020-01-07
#> 199 smben1421         Hosp     2020-01-07              Glinifera 2020-01-07
#> 200 smben1421         Hosp     2020-01-07          Sulfonylureas 2020-01-07
#> 201 smben1421         Hosp     2020-01-07                Statins 2020-01-07
#> 202 smben1421         Hosp     2020-01-07         Anticoagulants 2020-01-07
#> 203 smben1421         Hosp     2020-01-07             Biguanides 2020-01-07
#> 204 smben1421         Hosp     2020-01-07 Lipid-regulating drugs 2020-01-07
#> 205 smben1421         Hosp     2020-01-07 antiplatelet medicines 2020-01-07
#> 206 kuazo2021         outp     2020-01-16 antiplatelet medicines 2020-01-16
#> 207 kuazo2021         outp     2020-01-16                insulin 2020-01-16
#> 208 kuazo2021         outp     2020-01-16             Biguanides 2020-01-16
#> 209 kuazo2021         outp     2020-01-16                Statins 2020-01-16
#> 210 kuazo2021         outp     2020-01-16          beta-blockers 2020-01-16
#> 211 pmcrk4713         Hosp     2020-03-24                   ARBs 2020-03-24
#> 212 pmcrk4713         Hosp     2020-03-24          Sulfonylureas 2020-03-24
#> 213 pmcrk4713         Hosp     2020-03-24             Biguanides 2020-03-24
#> 214 pmcrk4713         Hosp     2020-03-24                insulin 2020-03-24
#> 215 pmcrk4713         Hosp     2020-03-24              Glinifera 2020-03-24
#> 216 pmcrk4713         Hosp     2020-03-24                Statins 2020-03-24
#> 217 pmcrk4713         Hosp     2020-03-24 Lipid-regulating drugs 2020-03-24
#> 218 pmcrk4713         Hosp     2020-03-24 antiplatelet medicines 2020-03-24
#> 219 xpixr3135         Hosp     2020-04-26                insulin 2020-04-26
#> 220 xpixr3135         Hosp     2020-04-26 Lipid-regulating drugs 2020-04-26
#> 221 xpixr3135         Hosp     2020-04-26         Anticoagulants 2020-04-26
#> 222 xpixr3135         Hosp     2020-04-26                   ARBs 2020-04-26
#> 223 xpixr3135         Hosp     2020-04-26              Glinifera 2020-04-26
#> 224 xpixr3135         Hosp     2020-04-26 antiplatelet medicines 2020-04-26
#> 225 twdrf4539         outp     2020-05-18             Biguanides 2020-05-18
#> 226 twdrf4539         outp     2020-05-18          Sulfonylureas 2020-05-18
#> 227 twdrf4539         outp     2020-05-18                Statins 2020-05-18
#> 228 twdrf4539         outp     2020-05-18              Glinifera 2020-05-18
#> 229 twdrf4539         outp     2020-05-18 Lipid-regulating drugs 2020-05-18
#> 230 twdrf4539         outp     2020-05-18         Anticoagulants 2020-05-18
#> 231 twdrf4539         outp     2020-05-18          beta-blockers 2020-05-18
#> 232 twdrf4539         outp     2020-05-18                insulin 2020-05-18
#> 233 twdrf4539         outp     2020-05-18 antiplatelet medicines 2020-05-18
#> 234 twdrf4539         outp     2020-05-18                   ARBs 2020-05-18
#> 235 svjhc1783         outp     2020-06-04 antiplatelet medicines 2020-06-04
#> 236 svjhc1783         outp     2020-06-04              Glinifera 2020-06-04
#> 237 svjhc1783         outp     2020-06-04         Anticoagulants 2020-06-04
#> 238 djiue5325         outp     2020-06-25 Lipid-regulating drugs 2020-06-25
#> 239 djiue5325         outp     2020-06-25          beta-blockers 2020-06-25
#> 240 djiue5325         outp     2020-06-25         Anticoagulants 2020-06-25
#> 241 djiue5325         outp     2020-06-25             Biguanides 2020-06-25
#> 242 djiue5325         outp     2020-06-25                insulin 2020-06-25
#> 243 djiue5325         outp     2020-06-25 antiplatelet medicines 2020-06-25
#> 244 djiue5325         outp     2020-06-25          Sulfonylureas 2020-06-25
#> 245 djiue5325         outp     2020-06-25                Statins 2020-06-25
#> 246 djiue5325         outp     2020-06-25              Glinifera 2020-06-25
#> 247 djiue5325         outp     2020-06-25                   ARBs 2020-06-25
#> 248 cknsq2879         Hosp     2020-07-18              Glinifera 2020-07-18
#> 249 cknsq2879         Hosp     2020-07-18          Sulfonylureas 2020-07-18
#> 250 cknsq2879         Hosp     2020-07-18 Lipid-regulating drugs 2020-07-18
#> 251 cknsq2879         Hosp     2020-07-18                Statins 2020-07-18
#> 252 cknsq2879         Hosp     2020-07-18          beta-blockers 2020-07-18
#> 253 tkpft4981         Hosp     2020-10-25          Sulfonylureas 2020-10-25
#> 254 tkpft4981         Hosp     2020-10-25                insulin 2020-10-25
#> 255 tkpft4981         Hosp     2020-10-25              Glinifera 2020-10-25
#> 256 tkpft4981         Hosp     2020-10-25 Lipid-regulating drugs 2020-10-25
#> 257 tkpft4981         Hosp     2020-10-25          beta-blockers 2020-10-25
#> 258 tkpft4981         Hosp     2020-10-25         Anticoagulants 2020-10-25
#> 259 tkpft4981         Hosp     2020-10-25             Biguanides 2020-10-25
#> 260 tkpft4981         Hosp     2020-10-25                Statins 2020-10-25
#> 
#> $lab
#>      visit_id patient_type admission_date lab_name lab_va  test_date
#> 1   fnhjl5940         Hosp     2014-01-22      LDL      2 2014-01-23
#> 2   fnhjl5940         Hosp     2014-01-22      WBC      4 2014-01-23
#> 3   fnhjl5940         Hosp     2014-01-22       TG     13 2014-01-23
#> 4   fnhjl5940         Hosp     2014-01-22      HDL      2 2014-01-23
#> 5   fnhjl5940         Hosp     2014-01-22       TC      5 2014-01-23
#> 6   ieetc5356         Hosp     2014-02-14    HbA1c      4 2014-02-15
#> 7   ieetc5356         Hosp     2014-02-14       TG     18 2014-02-15
#> 8   ieetc5356         Hosp     2014-02-14       TC      3 2014-02-15
#> 9   ieetc5356         Hosp     2014-02-14      HDL      4 2014-02-15
#> 10  ieetc5356         Hosp     2014-02-14      LDL      1 2014-02-15
#> 11  ieetc5356         Hosp     2014-02-14      WBC     11 2014-02-15
#> 12  svpiu3306         Hosp     2014-07-07    HbA1c      3 2014-07-08
#> 13  svpiu3306         Hosp     2014-07-07       TG     16 2014-07-08
#> 14  svpiu3306         Hosp     2014-07-07      HDL      4 2014-07-08
#> 15  ytyga3981         outp     2014-10-18      LDL      4 2014-10-19
#> 16  ytyga3981         outp     2014-10-18       TC      6 2014-10-19
#> 17  ytyga3981         outp     2014-10-18      HDL      2 2014-10-19
#> 18  ytyga3981         outp     2014-10-18       TG      9 2014-10-19
#> 19  ytyga3981         outp     2014-10-18      WBC     14 2014-10-19
#> 20  tntmb2887         Hosp     2014-11-27       TC      4 2014-11-28
#> 21  tntmb2887         Hosp     2014-11-27       TG      9 2014-11-28
#> 22  tntmb2887         Hosp     2014-11-27    HbA1c      7 2014-11-28
#> 23  tntmb2887         Hosp     2014-11-27      WBC      3 2014-11-28
#> 24  qnoqv3107         outp     2015-01-25       TG      1 2015-01-26
#> 25  qnoqv3107         outp     2015-01-25       TC      1 2015-01-26
#> 26  qnoqv3107         outp     2015-01-25      WBC     10 2015-01-26
#> 27  fmsbq3634         Hosp     2015-01-26      HDL      5 2015-01-27
#> 28  fmsbq3634         Hosp     2015-01-26       TC      4 2015-01-27
#> 29  fmsbq3634         Hosp     2015-01-26    HbA1c      3 2015-01-27
#> 30  fheez3668         Hosp     2015-05-20       TC      4 2015-05-21
#> 31  fheez3668         Hosp     2015-05-20      WBC      8 2015-05-21
#> 32  fheez3668         Hosp     2015-05-20       TG      2 2015-05-21
#> 33  fheez3668         Hosp     2015-05-20    HbA1c      2 2015-05-21
#> 34  fheez3668         Hosp     2015-05-20      HDL      3 2015-05-21
#> 35  qrcql3396         outp     2015-06-27      LDL      2 2015-06-28
#> 36  qrcql3396         outp     2015-06-27      WBC     10 2015-06-28
#> 37  qrcql3396         outp     2015-06-27      HDL      2 2015-06-28
#> 38  qrcql3396         outp     2015-06-27    HbA1c      1 2015-06-28
#> 39  illqe2480         Hosp     2016-01-26      LDL      2 2016-01-27
#> 40  illqe2480         Hosp     2016-01-26       TG      6 2016-01-27
#> 41  illqe2480         Hosp     2016-01-26    HbA1c      5 2016-01-27
#> 42  illqe2480         Hosp     2016-01-26      WBC     12 2016-01-27
#> 43  illqe2480         Hosp     2016-01-26      HDL      1 2016-01-27
#> 44  illqe2480         Hosp     2016-01-26       TC      7 2016-01-27
#> 45  wczwb4690         outp     2016-03-08    HbA1c      2 2016-03-09
#> 46  wczwb4690         outp     2016-03-08       TG     15 2016-03-09
#> 47  wczwb4690         outp     2016-03-08      HDL      4 2016-03-09
#> 48  hfopx3926         Hosp     2016-03-27      HDL      4 2016-03-28
#> 49  hfopx3926         Hosp     2016-03-27      WBC     14 2016-03-28
#> 50  hfopx3926         Hosp     2016-03-27      LDL      2 2016-03-28
#> 51  hfopx3926         Hosp     2016-03-27       TG      8 2016-03-28
#> 52  klzyf4532         Hosp     2016-05-31      LDL      2 2016-06-01
#> 53  klzyf4532         Hosp     2016-05-31      HDL      5 2016-06-01
#> 54  klzyf4532         Hosp     2016-05-31      WBC     13 2016-06-01
#> 55  klzyf4532         Hosp     2016-05-31    HbA1c      4 2016-06-01
#> 56  ojava5964         Hosp     2016-08-13       TC      4 2016-08-14
#> 57  ojava5964         Hosp     2016-08-13      HDL      2 2016-08-14
#> 58  ojava5964         Hosp     2016-08-13       TG      6 2016-08-14
#> 59  ojava5964         Hosp     2016-08-13    HbA1c      5 2016-08-14
#> 60  skmtg5971         outp     2017-01-17      HDL      3 2017-01-18
#> 61  skmtg5971         outp     2017-01-17      WBC     16 2017-01-18
#> 62  skmtg5971         outp     2017-01-17       TC      2 2017-01-18
#> 63  skmtg5971         outp     2017-01-17    HbA1c      9 2017-01-18
#> 64  skmtg5971         outp     2017-01-17      LDL      1 2017-01-18
#> 65  skmtg5971         outp     2017-01-17       TG      6 2017-01-18
#> 66  beklf3627         Hosp     2017-06-29    HbA1c      2 2017-06-30
#> 67  beklf3627         Hosp     2017-06-29       TG     17 2017-06-30
#> 68  beklf3627         Hosp     2017-06-29      LDL      5 2017-06-30
#> 69  beklf3627         Hosp     2017-06-29      WBC     12 2017-06-30
#> 70  beklf3627         Hosp     2017-06-29       TC      6 2017-06-30
#> 71  bvlns4639         outp     2017-07-02      WBC     11 2017-07-03
#> 72  bvlns4639         outp     2017-07-02      HDL      3 2017-07-03
#> 73  bvlns4639         outp     2017-07-02       TG     16 2017-07-03
#> 74  bvlns4639         outp     2017-07-02    HbA1c     11 2017-07-03
#> 75  dallj4117         Hosp     2017-07-21      WBC      6 2017-07-22
#> 76  dallj4117         Hosp     2017-07-21       TG      2 2017-07-22
#> 77  dallj4117         Hosp     2017-07-21      LDL      2 2017-07-22
#> 78  dallj4117         Hosp     2017-07-21      HDL      4 2017-07-22
#> 79  dallj4117         Hosp     2017-07-21       TC      6 2017-07-22
#> 80  egzws5800         Hosp     2017-09-02    HbA1c      1 2017-09-03
#> 81  egzws5800         Hosp     2017-09-02      HDL      1 2017-09-03
#> 82  egzws5800         Hosp     2017-09-02      LDL      5 2017-09-03
#> 83  egzws5800         Hosp     2017-09-02      WBC     13 2017-09-03
#> 84  wnlkn3812         outp     2017-10-31       TC      3 2017-11-01
#> 85  wnlkn3812         outp     2017-10-31      LDL      5 2017-11-01
#> 86  wnlkn3812         outp     2017-10-31      WBC      8 2017-11-01
#> 87  wnlkn3812         outp     2017-10-31       TG      2 2017-11-01
#> 88  hkwat1476         Hosp     2018-04-30    HbA1c      8 2018-05-01
#> 89  hkwat1476         Hosp     2018-04-30      WBC     10 2018-05-01
#> 90  hkwat1476         Hosp     2018-04-30      LDL      2 2018-05-01
#> 91  hkwat1476         Hosp     2018-04-30       TC      6 2018-05-01
#> 92  vrpfw3655         Hosp     2018-05-07       TC      5 2018-05-08
#> 93  vrpfw3655         Hosp     2018-05-07    HbA1c      7 2018-05-08
#> 94  vrpfw3655         Hosp     2018-05-07      LDL      3 2018-05-08
#> 95  vrpfw3655         Hosp     2018-05-07       TG     15 2018-05-08
#> 96  fbdhh2630         outp     2018-06-03       TC      7 2018-06-04
#> 97  fbdhh2630         outp     2018-06-03    HbA1c     10 2018-06-04
#> 98  fbdhh2630         outp     2018-06-03       TG      1 2018-06-04
#> 99  ilgvm4712         Hosp     2018-09-30      LDL      5 2018-10-01
#> 100 ilgvm4712         Hosp     2018-09-30      WBC     14 2018-10-01
#> 101 ilgvm4712         Hosp     2018-09-30      HDL      4 2018-10-01
#> 102 dfdtu5442         outp     2019-04-20       TC      2 2019-04-21
#> 103 dfdtu5442         outp     2019-04-20       TG      8 2019-04-21
#> 104 dfdtu5442         outp     2019-04-20    HbA1c      4 2019-04-21
#> 105 dfdtu5442         outp     2019-04-20      HDL      2 2019-04-21
#> 106 bfwzd1905         outp     2019-05-23       TG      4 2019-05-24
#> 107 bfwzd1905         outp     2019-05-23    HbA1c      4 2019-05-24
#> 108 bfwzd1905         outp     2019-05-23      WBC     17 2019-05-24
#> 109 pwsks4093         outp     2019-06-02      WBC     11 2019-06-03
#> 110 pwsks4093         outp     2019-06-02      HDL      4 2019-06-03
#> 111 pwsks4093         outp     2019-06-02      LDL      1 2019-06-03
#> 112 pjeto4451         outp     2019-07-20      HDL      4 2019-07-21
#> 113 pjeto4451         outp     2019-07-20      WBC     20 2019-07-21
#> 114 pjeto4451         outp     2019-07-20       TC      2 2019-07-21
#> 115 pjeto4451         outp     2019-07-20    HbA1c      7 2019-07-21
#> 116 zsfbm3006         outp     2019-10-18       TC      1 2019-10-19
#> 117 zsfbm3006         outp     2019-10-18    HbA1c      3 2019-10-19
#> 118 zsfbm3006         outp     2019-10-18       TG     11 2019-10-19
#> 119 zsfbm3006         outp     2019-10-18      HDL      4 2019-10-19
#> 120 gsutd5982         outp     2019-12-02      HDL      1 2019-12-03
#> 121 gsutd5982         outp     2019-12-02      LDL      2 2019-12-03
#> 122 gsutd5982         outp     2019-12-02      WBC      4 2019-12-03
#> 123 gsutd5982         outp     2019-12-02    HbA1c     10 2019-12-03
#> 124 gsutd5982         outp     2019-12-02       TG     13 2019-12-03
#> 125 qmwji3541         outp     2019-12-08       TC      4 2019-12-09
#> 126 qmwji3541         outp     2019-12-08      HDL      1 2019-12-09
#> 127 qmwji3541         outp     2019-12-08      LDL      1 2019-12-09
#> 128 qmwji3541         outp     2019-12-08       TG     17 2019-12-09
#> 129 qmwji3541         outp     2019-12-08      WBC     13 2019-12-09
#> 130 qmwji3541         outp     2019-12-08    HbA1c      8 2019-12-09
#> 131 smben1421         Hosp     2020-01-07      HDL      3 2020-01-08
#> 132 smben1421         Hosp     2020-01-07      LDL      1 2020-01-08
#> 133 smben1421         Hosp     2020-01-07       TC      4 2020-01-08
#> 134 smben1421         Hosp     2020-01-07      WBC      3 2020-01-08
#> 135 smben1421         Hosp     2020-01-07       TG      5 2020-01-08
#> 136 kuazo2021         outp     2020-01-16    HbA1c      6 2020-01-17
#> 137 kuazo2021         outp     2020-01-16       TG     14 2020-01-17
#> 138 kuazo2021         outp     2020-01-16       TC      4 2020-01-17
#> 139 kuazo2021         outp     2020-01-16      WBC      7 2020-01-17
#> 140 kuazo2021         outp     2020-01-16      HDL      5 2020-01-17
#> 141 pmcrk4713         Hosp     2020-03-24      HDL      5 2020-03-25
#> 142 pmcrk4713         Hosp     2020-03-24       TC      2 2020-03-25
#> 143 pmcrk4713         Hosp     2020-03-24      WBC      3 2020-03-25
#> 144 xpixr3135         Hosp     2020-04-26    HbA1c     12 2020-04-27
#> 145 xpixr3135         Hosp     2020-04-26       TG     12 2020-04-27
#> 146 xpixr3135         Hosp     2020-04-26       TC      4 2020-04-27
#> 147 xpixr3135         Hosp     2020-04-26      LDL      2 2020-04-27
#> 148 xpixr3135         Hosp     2020-04-26      HDL      2 2020-04-27
#> 149 twdrf4539         outp     2020-05-18      WBC     19 2020-05-19
#> 150 twdrf4539         outp     2020-05-18       TC      3 2020-05-19
#> 151 twdrf4539         outp     2020-05-18    HbA1c      3 2020-05-19
#> 152 twdrf4539         outp     2020-05-18      HDL      4 2020-05-19
#> 153 twdrf4539         outp     2020-05-18      LDL      5 2020-05-19
#> 154 twdrf4539         outp     2020-05-18       TG     19 2020-05-19
#> 155 svjhc1783         outp     2020-06-04      LDL      4 2020-06-05
#> 156 svjhc1783         outp     2020-06-04      HDL      3 2020-06-05
#> 157 svjhc1783         outp     2020-06-04       TG     18 2020-06-05
#> 158 svjhc1783         outp     2020-06-04    HbA1c      2 2020-06-05
#> 159 svjhc1783         outp     2020-06-04       TC      6 2020-06-05
#> 160 djiue5325         outp     2020-06-25       TC      2 2020-06-26
#> 161 djiue5325         outp     2020-06-25      WBC     11 2020-06-26
#> 162 djiue5325         outp     2020-06-25       TG      8 2020-06-26
#> 163 cknsq2879         Hosp     2020-07-18       TG     15 2020-07-19
#> 164 cknsq2879         Hosp     2020-07-18      HDL      4 2020-07-19
#> 165 cknsq2879         Hosp     2020-07-18      WBC     20 2020-07-19
#> 166 tkpft4981         Hosp     2020-10-25      WBC      3 2020-10-26
#> 167 tkpft4981         Hosp     2020-10-25      LDL      1 2020-10-26
#> 168 tkpft4981         Hosp     2020-10-25      HDL      2 2020-10-26
#> 169 tkpft4981         Hosp     2020-10-25       TC      7 2020-10-26
#> 170 tkpft4981         Hosp     2020-10-25    HbA1c      7 2020-10-26
#> 
#> $diag
#> # A tibble: 40 × 5
#>    visit_id  patient_type admission_date diag_date  diag_union                  
#>    <chr>     <chr>        <date>         <date>     <chr>                       
#>  1 fnhjl5940 Hosp         2014-01-22     2014-01-23 Lymphatic carcinoma,Periphe…
#>  2 ieetc5356 Hosp         2014-02-14     2014-02-15 diabetes,Peptic ulcers,Stro…
#>  3 svpiu3306 Hosp         2014-07-07     2014-07-08 Stroke,diabetes,Dyslipidemi…
#>  4 ytyga3981 outp         2014-10-18     2014-10-19 Symptomatic neuropathy,Lymp…
#>  5 tntmb2887 Hosp         2014-11-27     2014-11-28 NA,Lymphatic carcinoma,Peri…
#>  6 qnoqv3107 outp         2015-01-25     2015-01-26 NA                          
#>  7 fmsbq3634 Hosp         2015-01-26     2015-01-27 NA,Peripheral vascular dise…
#>  8 fheez3668 Hosp         2015-05-20     2015-05-21 NA,Dyslipidemia,Chronic lun…
#>  9 qrcql3396 outp         2015-06-27     2015-06-28 Peripheral vascular disease…
#> 10 illqe2480 Hosp         2016-01-26     2016-01-27 Moderate and severe nephrop…
#> # … with 30 more rows
#>