calculate the virtual dataset of a patient

calculate the virtual dataset of a patient

Usage

Virtual_med_one(
  birth_start_date,
  birth_end_date,
  study_start_date,
  study_end_date,
  average_year_visit_n
)

Arguments

birth_start_date

Minimum expected birth date of the patient.

birth_end_date

Maximum expected birth date of the patient.

study_start_date

The starting time of the baseline period specified in the study.

study_end_date

The end time of the follow-up period specified in the study

average_year_visit_n

The average number of visits by patients each year is 6 by default. The data are from the statistical records of the Chinese people.

Value

Returns a data set of a patients, including a doctor's table, prescription table, diagnosis table, and biochemical test table.

Examples

Virtual_med_one(
 birth_start_date = "1980-01-01",birth_end_date ="2000-12-31" ,
   study_start_date = "2014-01-01",study_end_date ="2020-12-31",
   average_year_visit_n = 6)
#> $pat
#>     visit_id    sex  birthdate patient_type admission_date Discharge_date
#> 1  kzjjh5204 female 1993-11-17         outp     2014-01-25     2014-02-06
#> 2  rzkxe3488 female 1993-11-17         Hosp     2014-02-17     2014-03-01
#> 3  bdyxs1183 female 1993-11-17         outp     2014-03-01     2014-03-13
#> 4  iewtg2829 female 1993-11-17         outp     2015-01-08     2015-01-20
#> 5  owyfz5906 female 1993-11-17         outp     2015-01-20     2015-02-01
#> 6  qqgqp2762 female 1993-11-17         Hosp     2015-05-09     2015-05-21
#> 7  almth4311 female 1993-11-17         outp     2015-06-02     2015-06-14
#> 8  zyace1484 female 1993-11-17         outp     2015-12-24     2016-01-05
#> 9  gjxhs4505 female 1993-11-17         outp     2016-04-04     2016-04-16
#> 10 sqjml3858 female 1993-11-17         Hosp     2016-09-22     2016-10-04
#> 11 zphxl3698 female 1993-11-17         Hosp     2017-03-01     2017-03-13
#> 12 qvoob3019 female 1993-11-17         Hosp     2018-05-10     2018-05-22
#> 13 jtffi5455 female 1993-11-17         Hosp     2018-06-14     2018-06-26
#> 14 ujibc4520 female 1993-11-17         Hosp     2018-08-17     2018-08-29
#> 15 lrjyn2283 female 1993-11-17         Hosp     2018-10-31     2018-11-12
#> 16 ahieg5153 female 1993-11-17         Hosp     2018-12-28     2019-01-09
#> 17 eoejs3255 female 1993-11-17         Hosp     2019-02-10     2019-02-22
#> 18 oqvdu1681 female 1993-11-17         Hosp     2019-04-27     2019-05-09
#> 19 izmao1468 female 1993-11-17         outp     2019-11-25     2019-12-07
#> 20 gebyo5184 female 1993-11-17         outp     2019-12-16     2019-12-28
#> 21 ibjnh1074 female 1993-11-17         outp     2020-03-14     2020-03-26
#> 22 ubmxe4497 female 1993-11-17         outp     2020-11-11     2020-11-23
#> 
#> $prec
#>      visit_id patient_type admission_date              prec_name  prec_date
#> 1   kzjjh5204         outp     2014-01-25              Glinifera 2014-01-25
#> 2   kzjjh5204         outp     2014-01-25          Sulfonylureas 2014-01-25
#> 3   kzjjh5204         outp     2014-01-25                insulin 2014-01-25
#> 4   kzjjh5204         outp     2014-01-25                Statins 2014-01-25
#> 5   rzkxe3488         Hosp     2014-02-17                   ARBs 2014-02-17
#> 6   rzkxe3488         Hosp     2014-02-17                insulin 2014-02-17
#> 7   rzkxe3488         Hosp     2014-02-17 Lipid-regulating drugs 2014-02-17
#> 8   rzkxe3488         Hosp     2014-02-17 antiplatelet medicines 2014-02-17
#> 9   rzkxe3488         Hosp     2014-02-17              Glinifera 2014-02-17
#> 10  rzkxe3488         Hosp     2014-02-17                Statins 2014-02-17
#> 11  bdyxs1183         outp     2014-03-01              Glinifera 2014-03-01
#> 12  bdyxs1183         outp     2014-03-01 Lipid-regulating drugs 2014-03-01
#> 13  bdyxs1183         outp     2014-03-01                   ARBs 2014-03-01
#> 14  iewtg2829         outp     2015-01-08 Lipid-regulating drugs 2015-01-08
#> 15  iewtg2829         outp     2015-01-08                insulin 2015-01-08
#> 16  iewtg2829         outp     2015-01-08          Sulfonylureas 2015-01-08
#> 17  iewtg2829         outp     2015-01-08              Glinifera 2015-01-08
#> 18  owyfz5906         outp     2015-01-20          Sulfonylureas 2015-01-20
#> 19  owyfz5906         outp     2015-01-20         Anticoagulants 2015-01-20
#> 20  owyfz5906         outp     2015-01-20              Glinifera 2015-01-20
#> 21  owyfz5906         outp     2015-01-20             Biguanides 2015-01-20
#> 22  owyfz5906         outp     2015-01-20                   ARBs 2015-01-20
#> 23  owyfz5906         outp     2015-01-20 Lipid-regulating drugs 2015-01-20
#> 24  owyfz5906         outp     2015-01-20                insulin 2015-01-20
#> 25  owyfz5906         outp     2015-01-20 antiplatelet medicines 2015-01-20
#> 26  qqgqp2762         Hosp     2015-05-09                Statins 2015-05-09
#> 27  qqgqp2762         Hosp     2015-05-09          beta-blockers 2015-05-09
#> 28  qqgqp2762         Hosp     2015-05-09                insulin 2015-05-09
#> 29  qqgqp2762         Hosp     2015-05-09                   ARBs 2015-05-09
#> 30  almth4311         outp     2015-06-02              Glinifera 2015-06-02
#> 31  almth4311         outp     2015-06-02             Biguanides 2015-06-02
#> 32  almth4311         outp     2015-06-02          beta-blockers 2015-06-02
#> 33  almth4311         outp     2015-06-02                Statins 2015-06-02
#> 34  almth4311         outp     2015-06-02 antiplatelet medicines 2015-06-02
#> 35  almth4311         outp     2015-06-02 Lipid-regulating drugs 2015-06-02
#> 36  almth4311         outp     2015-06-02          Sulfonylureas 2015-06-02
#> 37  almth4311         outp     2015-06-02         Anticoagulants 2015-06-02
#> 38  almth4311         outp     2015-06-02                insulin 2015-06-02
#> 39  zyace1484         outp     2015-12-24          beta-blockers 2015-12-24
#> 40  zyace1484         outp     2015-12-24              Glinifera 2015-12-24
#> 41  zyace1484         outp     2015-12-24                insulin 2015-12-24
#> 42  zyace1484         outp     2015-12-24                   ARBs 2015-12-24
#> 43  gjxhs4505         outp     2016-04-04          beta-blockers 2016-04-04
#> 44  gjxhs4505         outp     2016-04-04                Statins 2016-04-04
#> 45  gjxhs4505         outp     2016-04-04                   ARBs 2016-04-04
#> 46  gjxhs4505         outp     2016-04-04          Sulfonylureas 2016-04-04
#> 47  gjxhs4505         outp     2016-04-04                insulin 2016-04-04
#> 48  gjxhs4505         outp     2016-04-04         Anticoagulants 2016-04-04
#> 49  sqjml3858         Hosp     2016-09-22 antiplatelet medicines 2016-09-22
#> 50  sqjml3858         Hosp     2016-09-22          beta-blockers 2016-09-22
#> 51  sqjml3858         Hosp     2016-09-22         Anticoagulants 2016-09-22
#> 52  sqjml3858         Hosp     2016-09-22          Sulfonylureas 2016-09-22
#> 53  zphxl3698         Hosp     2017-03-01                insulin 2017-03-01
#> 54  zphxl3698         Hosp     2017-03-01 Lipid-regulating drugs 2017-03-01
#> 55  zphxl3698         Hosp     2017-03-01          Sulfonylureas 2017-03-01
#> 56  qvoob3019         Hosp     2018-05-10          Sulfonylureas 2018-05-10
#> 57  qvoob3019         Hosp     2018-05-10          beta-blockers 2018-05-10
#> 58  qvoob3019         Hosp     2018-05-10         Anticoagulants 2018-05-10
#> 59  qvoob3019         Hosp     2018-05-10 antiplatelet medicines 2018-05-10
#> 60  qvoob3019         Hosp     2018-05-10             Biguanides 2018-05-10
#> 61  qvoob3019         Hosp     2018-05-10              Glinifera 2018-05-10
#> 62  qvoob3019         Hosp     2018-05-10 Lipid-regulating drugs 2018-05-10
#> 63  jtffi5455         Hosp     2018-06-14          Sulfonylureas 2018-06-14
#> 64  jtffi5455         Hosp     2018-06-14                Statins 2018-06-14
#> 65  jtffi5455         Hosp     2018-06-14 Lipid-regulating drugs 2018-06-14
#> 66  ujibc4520         Hosp     2018-08-17                   ARBs 2018-08-17
#> 67  ujibc4520         Hosp     2018-08-17 antiplatelet medicines 2018-08-17
#> 68  ujibc4520         Hosp     2018-08-17                insulin 2018-08-17
#> 69  ujibc4520         Hosp     2018-08-17              Glinifera 2018-08-17
#> 70  ujibc4520         Hosp     2018-08-17         Anticoagulants 2018-08-17
#> 71  lrjyn2283         Hosp     2018-10-31                insulin 2018-10-31
#> 72  lrjyn2283         Hosp     2018-10-31                   ARBs 2018-10-31
#> 73  lrjyn2283         Hosp     2018-10-31         Anticoagulants 2018-10-31
#> 74  lrjyn2283         Hosp     2018-10-31 Lipid-regulating drugs 2018-10-31
#> 75  lrjyn2283         Hosp     2018-10-31                Statins 2018-10-31
#> 76  lrjyn2283         Hosp     2018-10-31              Glinifera 2018-10-31
#> 77  lrjyn2283         Hosp     2018-10-31          beta-blockers 2018-10-31
#> 78  lrjyn2283         Hosp     2018-10-31          Sulfonylureas 2018-10-31
#> 79  lrjyn2283         Hosp     2018-10-31             Biguanides 2018-10-31
#> 80  lrjyn2283         Hosp     2018-10-31 antiplatelet medicines 2018-10-31
#> 81  ahieg5153         Hosp     2018-12-28          Sulfonylureas 2018-12-28
#> 82  ahieg5153         Hosp     2018-12-28         Anticoagulants 2018-12-28
#> 83  ahieg5153         Hosp     2018-12-28             Biguanides 2018-12-28
#> 84  ahieg5153         Hosp     2018-12-28 Lipid-regulating drugs 2018-12-28
#> 85  ahieg5153         Hosp     2018-12-28              Glinifera 2018-12-28
#> 86  ahieg5153         Hosp     2018-12-28                insulin 2018-12-28
#> 87  ahieg5153         Hosp     2018-12-28 antiplatelet medicines 2018-12-28
#> 88  ahieg5153         Hosp     2018-12-28                   ARBs 2018-12-28
#> 89  ahieg5153         Hosp     2018-12-28          beta-blockers 2018-12-28
#> 90  ahieg5153         Hosp     2018-12-28                Statins 2018-12-28
#> 91  eoejs3255         Hosp     2019-02-10 antiplatelet medicines 2019-02-10
#> 92  eoejs3255         Hosp     2019-02-10                insulin 2019-02-10
#> 93  eoejs3255         Hosp     2019-02-10              Glinifera 2019-02-10
#> 94  eoejs3255         Hosp     2019-02-10                Statins 2019-02-10
#> 95  oqvdu1681         Hosp     2019-04-27              Glinifera 2019-04-27
#> 96  oqvdu1681         Hosp     2019-04-27         Anticoagulants 2019-04-27
#> 97  oqvdu1681         Hosp     2019-04-27          beta-blockers 2019-04-27
#> 98  oqvdu1681         Hosp     2019-04-27                   ARBs 2019-04-27
#> 99  oqvdu1681         Hosp     2019-04-27 antiplatelet medicines 2019-04-27
#> 100 izmao1468         outp     2019-11-25         Anticoagulants 2019-11-25
#> 101 izmao1468         outp     2019-11-25                   ARBs 2019-11-25
#> 102 izmao1468         outp     2019-11-25                insulin 2019-11-25
#> 103 izmao1468         outp     2019-11-25 Lipid-regulating drugs 2019-11-25
#> 104 izmao1468         outp     2019-11-25                Statins 2019-11-25
#> 105 izmao1468         outp     2019-11-25 antiplatelet medicines 2019-11-25
#> 106 izmao1468         outp     2019-11-25          Sulfonylureas 2019-11-25
#> 107 izmao1468         outp     2019-11-25             Biguanides 2019-11-25
#> 108 gebyo5184         outp     2019-12-16              Glinifera 2019-12-16
#> 109 gebyo5184         outp     2019-12-16                insulin 2019-12-16
#> 110 gebyo5184         outp     2019-12-16             Biguanides 2019-12-16
#> 111 gebyo5184         outp     2019-12-16                   ARBs 2019-12-16
#> 112 gebyo5184         outp     2019-12-16 antiplatelet medicines 2019-12-16
#> 113 gebyo5184         outp     2019-12-16          beta-blockers 2019-12-16
#> 114 gebyo5184         outp     2019-12-16          Sulfonylureas 2019-12-16
#> 115 gebyo5184         outp     2019-12-16 Lipid-regulating drugs 2019-12-16
#> 116 ibjnh1074         outp     2020-03-14 antiplatelet medicines 2020-03-14
#> 117 ibjnh1074         outp     2020-03-14                   ARBs 2020-03-14
#> 118 ibjnh1074         outp     2020-03-14          beta-blockers 2020-03-14
#> 119 ibjnh1074         outp     2020-03-14 Lipid-regulating drugs 2020-03-14
#> 120 ibjnh1074         outp     2020-03-14              Glinifera 2020-03-14
#> 121 ibjnh1074         outp     2020-03-14             Biguanides 2020-03-14
#> 122 ibjnh1074         outp     2020-03-14                Statins 2020-03-14
#> 123 ibjnh1074         outp     2020-03-14         Anticoagulants 2020-03-14
#> 124 ibjnh1074         outp     2020-03-14          Sulfonylureas 2020-03-14
#> 125 ibjnh1074         outp     2020-03-14                insulin 2020-03-14
#> 126 ubmxe4497         outp     2020-11-11                   ARBs 2020-11-11
#> 127 ubmxe4497         outp     2020-11-11         Anticoagulants 2020-11-11
#> 128 ubmxe4497         outp     2020-11-11 Lipid-regulating drugs 2020-11-11
#> 129 ubmxe4497         outp     2020-11-11                insulin 2020-11-11
#> 130 ubmxe4497         outp     2020-11-11              Glinifera 2020-11-11
#> 131 ubmxe4497         outp     2020-11-11             Biguanides 2020-11-11
#> 132 ubmxe4497         outp     2020-11-11          Sulfonylureas 2020-11-11
#> 133 ubmxe4497         outp     2020-11-11          beta-blockers 2020-11-11
#> 134 ubmxe4497         outp     2020-11-11                Statins 2020-11-11
#> 
#> $lab
#>      visit_id patient_type admission_date lab_name lab_va  test_date
#> 1   kzjjh5204         outp     2014-01-25      HDL      5 2014-01-26
#> 2   kzjjh5204         outp     2014-01-25      WBC      8 2014-01-26
#> 3   kzjjh5204         outp     2014-01-25    HbA1c      7 2014-01-26
#> 4   kzjjh5204         outp     2014-01-25      LDL      2 2014-01-26
#> 5   kzjjh5204         outp     2014-01-25       TC      3 2014-01-26
#> 6   rzkxe3488         Hosp     2014-02-17       TG     10 2014-02-18
#> 7   rzkxe3488         Hosp     2014-02-17       TC      4 2014-02-18
#> 8   rzkxe3488         Hosp     2014-02-17      WBC     10 2014-02-18
#> 9   rzkxe3488         Hosp     2014-02-17      HDL      1 2014-02-18
#> 10  rzkxe3488         Hosp     2014-02-17    HbA1c     12 2014-02-18
#> 11  rzkxe3488         Hosp     2014-02-17      LDL      3 2014-02-18
#> 12  bdyxs1183         outp     2014-03-01       TG      2 2014-03-02
#> 13  bdyxs1183         outp     2014-03-01      WBC      7 2014-03-02
#> 14  bdyxs1183         outp     2014-03-01      HDL      4 2014-03-02
#> 15  bdyxs1183         outp     2014-03-01       TC      6 2014-03-02
#> 16  bdyxs1183         outp     2014-03-01    HbA1c      1 2014-03-02
#> 17  bdyxs1183         outp     2014-03-01      LDL      3 2014-03-02
#> 18  iewtg2829         outp     2015-01-08      LDL      2 2015-01-09
#> 19  iewtg2829         outp     2015-01-08    HbA1c      6 2015-01-09
#> 20  iewtg2829         outp     2015-01-08      WBC      5 2015-01-09
#> 21  iewtg2829         outp     2015-01-08       TG     11 2015-01-09
#> 22  owyfz5906         outp     2015-01-20      HDL      1 2015-01-21
#> 23  owyfz5906         outp     2015-01-20    HbA1c      2 2015-01-21
#> 24  owyfz5906         outp     2015-01-20      WBC      1 2015-01-21
#> 25  owyfz5906         outp     2015-01-20       TC      3 2015-01-21
#> 26  owyfz5906         outp     2015-01-20      LDL      2 2015-01-21
#> 27  owyfz5906         outp     2015-01-20       TG      6 2015-01-21
#> 28  qqgqp2762         Hosp     2015-05-09       TG      7 2015-05-10
#> 29  qqgqp2762         Hosp     2015-05-09      HDL      4 2015-05-10
#> 30  qqgqp2762         Hosp     2015-05-09       TC      7 2015-05-10
#> 31  qqgqp2762         Hosp     2015-05-09      LDL      3 2015-05-10
#> 32  qqgqp2762         Hosp     2015-05-09      WBC      8 2015-05-10
#> 33  almth4311         outp     2015-06-02       TG     17 2015-06-03
#> 34  almth4311         outp     2015-06-02      HDL      5 2015-06-03
#> 35  almth4311         outp     2015-06-02    HbA1c     10 2015-06-03
#> 36  almth4311         outp     2015-06-02      LDL      5 2015-06-03
#> 37  zyace1484         outp     2015-12-24       TG     16 2015-12-25
#> 38  zyace1484         outp     2015-12-24      WBC     17 2015-12-25
#> 39  zyace1484         outp     2015-12-24    HbA1c      9 2015-12-25
#> 40  gjxhs4505         outp     2016-04-04       TC      1 2016-04-05
#> 41  gjxhs4505         outp     2016-04-04      WBC      7 2016-04-05
#> 42  gjxhs4505         outp     2016-04-04      HDL      5 2016-04-05
#> 43  gjxhs4505         outp     2016-04-04    HbA1c      2 2016-04-05
#> 44  gjxhs4505         outp     2016-04-04       TG      3 2016-04-05
#> 45  sqjml3858         Hosp     2016-09-22    HbA1c      2 2016-09-23
#> 46  sqjml3858         Hosp     2016-09-22       TC      5 2016-09-23
#> 47  sqjml3858         Hosp     2016-09-22       TG     20 2016-09-23
#> 48  sqjml3858         Hosp     2016-09-22      WBC     11 2016-09-23
#> 49  zphxl3698         Hosp     2017-03-01       TG      7 2017-03-02
#> 50  zphxl3698         Hosp     2017-03-01      LDL      1 2017-03-02
#> 51  zphxl3698         Hosp     2017-03-01    HbA1c      7 2017-03-02
#> 52  zphxl3698         Hosp     2017-03-01      WBC      4 2017-03-02
#> 53  qvoob3019         Hosp     2018-05-10    HbA1c      1 2018-05-11
#> 54  qvoob3019         Hosp     2018-05-10      HDL      2 2018-05-11
#> 55  qvoob3019         Hosp     2018-05-10       TC      5 2018-05-11
#> 56  qvoob3019         Hosp     2018-05-10      WBC      5 2018-05-11
#> 57  qvoob3019         Hosp     2018-05-10      LDL      1 2018-05-11
#> 58  jtffi5455         Hosp     2018-06-14      WBC     20 2018-06-15
#> 59  jtffi5455         Hosp     2018-06-14       TG      6 2018-06-15
#> 60  jtffi5455         Hosp     2018-06-14       TC      4 2018-06-15
#> 61  jtffi5455         Hosp     2018-06-14      HDL      2 2018-06-15
#> 62  ujibc4520         Hosp     2018-08-17    HbA1c      1 2018-08-18
#> 63  ujibc4520         Hosp     2018-08-17       TG      9 2018-08-18
#> 64  ujibc4520         Hosp     2018-08-17      LDL      3 2018-08-18
#> 65  lrjyn2283         Hosp     2018-10-31       TG     16 2018-11-01
#> 66  lrjyn2283         Hosp     2018-10-31      WBC      8 2018-11-01
#> 67  lrjyn2283         Hosp     2018-10-31       TC      5 2018-11-01
#> 68  lrjyn2283         Hosp     2018-10-31      LDL      2 2018-11-01
#> 69  ahieg5153         Hosp     2018-12-28    HbA1c      2 2018-12-29
#> 70  ahieg5153         Hosp     2018-12-28       TC      1 2018-12-29
#> 71  ahieg5153         Hosp     2018-12-28      WBC     12 2018-12-29
#> 72  ahieg5153         Hosp     2018-12-28      HDL      2 2018-12-29
#> 73  ahieg5153         Hosp     2018-12-28       TG     16 2018-12-29
#> 74  ahieg5153         Hosp     2018-12-28      LDL      3 2018-12-29
#> 75  eoejs3255         Hosp     2019-02-10       TG     20 2019-02-11
#> 76  eoejs3255         Hosp     2019-02-10      HDL      5 2019-02-11
#> 77  eoejs3255         Hosp     2019-02-10       TC      7 2019-02-11
#> 78  eoejs3255         Hosp     2019-02-10    HbA1c     11 2019-02-11
#> 79  eoejs3255         Hosp     2019-02-10      LDL      4 2019-02-11
#> 80  oqvdu1681         Hosp     2019-04-27      LDL      3 2019-04-28
#> 81  oqvdu1681         Hosp     2019-04-27    HbA1c     12 2019-04-28
#> 82  oqvdu1681         Hosp     2019-04-27      WBC      9 2019-04-28
#> 83  oqvdu1681         Hosp     2019-04-27      HDL      5 2019-04-28
#> 84  izmao1468         outp     2019-11-25       TC      4 2019-11-26
#> 85  izmao1468         outp     2019-11-25      WBC      5 2019-11-26
#> 86  izmao1468         outp     2019-11-25      HDL      1 2019-11-26
#> 87  izmao1468         outp     2019-11-25      LDL      4 2019-11-26
#> 88  izmao1468         outp     2019-11-25       TG     10 2019-11-26
#> 89  gebyo5184         outp     2019-12-16      LDL      2 2019-12-17
#> 90  gebyo5184         outp     2019-12-16    HbA1c      6 2019-12-17
#> 91  gebyo5184         outp     2019-12-16      HDL      2 2019-12-17
#> 92  gebyo5184         outp     2019-12-16       TC      6 2019-12-17
#> 93  ibjnh1074         outp     2020-03-14      LDL      3 2020-03-15
#> 94  ibjnh1074         outp     2020-03-14       TC      5 2020-03-15
#> 95  ibjnh1074         outp     2020-03-14    HbA1c      1 2020-03-15
#> 96  ibjnh1074         outp     2020-03-14      HDL      3 2020-03-15
#> 97  ibjnh1074         outp     2020-03-14       TG     12 2020-03-15
#> 98  ibjnh1074         outp     2020-03-14      WBC     20 2020-03-15
#> 99  ubmxe4497         outp     2020-11-11    HbA1c     11 2020-11-12
#> 100 ubmxe4497         outp     2020-11-11       TC      1 2020-11-12
#> 101 ubmxe4497         outp     2020-11-11      LDL      4 2020-11-12
#> 
#> $diag
#> # A tibble: 22 × 5
#>    visit_id  patient_type admission_date diag_date  diag_union                  
#>    <chr>     <chr>        <date>         <date>     <chr>                       
#>  1 kzjjh5204 outp         2014-01-25     2014-01-26 Peripheral vascular disease…
#>  2 rzkxe3488 Hosp         2014-02-17     2014-02-18 NA,myocardial infarction,De…
#>  3 bdyxs1183 outp         2014-03-01     2014-03-02 Dyslipidemia,Mild liver dis…
#>  4 iewtg2829 outp         2015-01-08     2015-01-09 Stroke,Lymphatic carcinoma,…
#>  5 owyfz5906 outp         2015-01-20     2015-01-21 NA,myocardial infarction,Pe…
#>  6 qqgqp2762 Hosp         2015-05-09     2015-05-10 Symptomatic neuropathy,Peri…
#>  7 almth4311 outp         2015-06-02     2015-06-03 NA,Stroke,Symptomatic neuro…
#>  8 zyace1484 outp         2015-12-24     2015-12-25 Mild liver disease,myocardi…
#>  9 gjxhs4505 outp         2016-04-04     2016-04-05 NA,high blood pressure      
#> 10 sqjml3858 Hosp         2016-09-22     2016-09-23 Symptomatic neuropathy,high…
#> # … with 12 more rows
#>