calculate the virtual dataset of a patient

calculate the virtual dataset of a patient

Usage

Virtual_med_one(
  birth_start_date,
  birth_end_date,
  study_start_date,
  study_end_date,
  average_year_visit_n
)

Arguments

birth_start_date

Minimum expected birth date of the patient.

birth_end_date

Maximum expected birth date of the patient.

study_start_date

The starting time of the baseline period specified in the study.

study_end_date

The end time of the follow-up period specified in the study

average_year_visit_n

The average number of visits by patients each year is 6 by default. The data are from the statistical records of the Chinese people.

Value

Returns a data set of a patients, including a doctor's table, prescription table, diagnosis table, and biochemical test table.

Examples

Virtual_med_one(
 birth_start_date = "1980-01-01",birth_end_date ="2000-12-31" ,
   study_start_date = "2014-01-01",study_end_date ="2020-12-31",
   average_year_visit_n = 6)
#> $pat
#>     visit_id    sex  birthdate patient_type admission_date Discharge_date
#> 1  zgvbb2931 female 1983-08-11         Hosp     2014-05-14     2014-06-20
#> 2  egclx5067 female 1983-08-11         Hosp     2014-06-23     2014-07-30
#> 3  tuhol1267 female 1983-08-11         outp     2015-02-26     2015-04-04
#> 4  yyiia3143 female 1983-08-11         Hosp     2015-05-24     2015-06-30
#> 5  wvvwk5077 female 1983-08-11         outp     2015-07-08     2015-08-14
#> 6  qimcg2445 female 1983-08-11         Hosp     2015-08-08     2015-09-14
#> 7  tnqqj2754 female 1983-08-11         Hosp     2015-11-12     2015-12-19
#> 8  odgxj5634 female 1983-08-11         Hosp     2016-02-11     2016-03-19
#> 9  kkkpf4171 female 1983-08-11         outp     2016-05-07     2016-06-13
#> 10 opzoz2779 female 1983-08-11         Hosp     2016-09-06     2016-10-13
#> 11 hwjjh4986 female 1983-08-11         outp     2016-09-18     2016-10-25
#> 12 cpqzj5218 female 1983-08-11         Hosp     2017-02-01     2017-03-10
#> 13 oiegk5170 female 1983-08-11         Hosp     2017-02-16     2017-03-25
#> 14 pprzx3026 female 1983-08-11         outp     2017-05-10     2017-06-16
#> 15 wuukc4654 female 1983-08-11         outp     2017-08-01     2017-09-07
#> 16 vlfya2039 female 1983-08-11         outp     2017-12-05     2018-01-11
#> 17 fwctd1785 female 1983-08-11         outp     2017-12-06     2018-01-12
#> 18 gbbug3812 female 1983-08-11         Hosp     2017-12-23     2018-01-29
#> 19 cpttc5403 female 1983-08-11         Hosp     2018-02-06     2018-03-15
#> 20 vfpeh2440 female 1983-08-11         outp     2018-06-22     2018-07-29
#> 21 qgauj2201 female 1983-08-11         outp     2018-07-31     2018-09-06
#> 22 fnlaz3245 female 1983-08-11         Hosp     2018-08-09     2018-09-15
#> 23 xqnri2590 female 1983-08-11         Hosp     2018-08-29     2018-10-05
#> 24 qdnjd5149 female 1983-08-11         Hosp     2018-09-12     2018-10-19
#> 25 rviko4185 female 1983-08-11         outp     2018-10-04     2018-11-10
#> 26 tmcsy4261 female 1983-08-11         Hosp     2018-10-07     2018-11-13
#> 27 ikveg4786 female 1983-08-11         Hosp     2018-12-12     2019-01-18
#> 28 sxklc2322 female 1983-08-11         Hosp     2019-01-05     2019-02-11
#> 29 hnfot2199 female 1983-08-11         Hosp     2019-02-16     2019-03-25
#> 30 crght3835 female 1983-08-11         outp     2019-04-18     2019-05-25
#> 31 rlhki2360 female 1983-08-11         outp     2020-04-16     2020-05-23
#> 32 oavnw2458 female 1983-08-11         outp     2020-06-08     2020-07-15
#> 
#> $prec
#>      visit_id patient_type admission_date              prec_name  prec_date
#> 1   zgvbb2931         Hosp     2014-05-14          beta-blockers 2014-05-14
#> 2   zgvbb2931         Hosp     2014-05-14             Biguanides 2014-05-14
#> 3   zgvbb2931         Hosp     2014-05-14                Statins 2014-05-14
#> 4   zgvbb2931         Hosp     2014-05-14                   ARBs 2014-05-14
#> 5   zgvbb2931         Hosp     2014-05-14              Glinifera 2014-05-14
#> 6   egclx5067         Hosp     2014-06-23          Sulfonylureas 2014-06-23
#> 7   egclx5067         Hosp     2014-06-23 antiplatelet medicines 2014-06-23
#> 8   egclx5067         Hosp     2014-06-23             Biguanides 2014-06-23
#> 9   egclx5067         Hosp     2014-06-23 Lipid-regulating drugs 2014-06-23
#> 10  egclx5067         Hosp     2014-06-23          beta-blockers 2014-06-23
#> 11  egclx5067         Hosp     2014-06-23              Glinifera 2014-06-23
#> 12  egclx5067         Hosp     2014-06-23         Anticoagulants 2014-06-23
#> 13  tuhol1267         outp     2015-02-26                Statins 2015-02-26
#> 14  tuhol1267         outp     2015-02-26                   ARBs 2015-02-26
#> 15  tuhol1267         outp     2015-02-26          beta-blockers 2015-02-26
#> 16  tuhol1267         outp     2015-02-26                insulin 2015-02-26
#> 17  tuhol1267         outp     2015-02-26          Sulfonylureas 2015-02-26
#> 18  tuhol1267         outp     2015-02-26 antiplatelet medicines 2015-02-26
#> 19  tuhol1267         outp     2015-02-26              Glinifera 2015-02-26
#> 20  tuhol1267         outp     2015-02-26 Lipid-regulating drugs 2015-02-26
#> 21  tuhol1267         outp     2015-02-26         Anticoagulants 2015-02-26
#> 22  tuhol1267         outp     2015-02-26             Biguanides 2015-02-26
#> 23  yyiia3143         Hosp     2015-05-24 Lipid-regulating drugs 2015-05-24
#> 24  yyiia3143         Hosp     2015-05-24             Biguanides 2015-05-24
#> 25  yyiia3143         Hosp     2015-05-24                insulin 2015-05-24
#> 26  yyiia3143         Hosp     2015-05-24          Sulfonylureas 2015-05-24
#> 27  yyiia3143         Hosp     2015-05-24                   ARBs 2015-05-24
#> 28  yyiia3143         Hosp     2015-05-24                Statins 2015-05-24
#> 29  yyiia3143         Hosp     2015-05-24          beta-blockers 2015-05-24
#> 30  yyiia3143         Hosp     2015-05-24              Glinifera 2015-05-24
#> 31  wvvwk5077         outp     2015-07-08              Glinifera 2015-07-08
#> 32  wvvwk5077         outp     2015-07-08 Lipid-regulating drugs 2015-07-08
#> 33  wvvwk5077         outp     2015-07-08                   ARBs 2015-07-08
#> 34  wvvwk5077         outp     2015-07-08                Statins 2015-07-08
#> 35  qimcg2445         Hosp     2015-08-08             Biguanides 2015-08-08
#> 36  qimcg2445         Hosp     2015-08-08                insulin 2015-08-08
#> 37  qimcg2445         Hosp     2015-08-08          beta-blockers 2015-08-08
#> 38  tnqqj2754         Hosp     2015-11-12         Anticoagulants 2015-11-12
#> 39  tnqqj2754         Hosp     2015-11-12              Glinifera 2015-11-12
#> 40  tnqqj2754         Hosp     2015-11-12          Sulfonylureas 2015-11-12
#> 41  tnqqj2754         Hosp     2015-11-12 antiplatelet medicines 2015-11-12
#> 42  tnqqj2754         Hosp     2015-11-12                Statins 2015-11-12
#> 43  tnqqj2754         Hosp     2015-11-12             Biguanides 2015-11-12
#> 44  tnqqj2754         Hosp     2015-11-12 Lipid-regulating drugs 2015-11-12
#> 45  tnqqj2754         Hosp     2015-11-12                   ARBs 2015-11-12
#> 46  tnqqj2754         Hosp     2015-11-12          beta-blockers 2015-11-12
#> 47  tnqqj2754         Hosp     2015-11-12                insulin 2015-11-12
#> 48  odgxj5634         Hosp     2016-02-11          beta-blockers 2016-02-11
#> 49  odgxj5634         Hosp     2016-02-11                   ARBs 2016-02-11
#> 50  odgxj5634         Hosp     2016-02-11             Biguanides 2016-02-11
#> 51  kkkpf4171         outp     2016-05-07                insulin 2016-05-07
#> 52  kkkpf4171         outp     2016-05-07             Biguanides 2016-05-07
#> 53  kkkpf4171         outp     2016-05-07 Lipid-regulating drugs 2016-05-07
#> 54  kkkpf4171         outp     2016-05-07                   ARBs 2016-05-07
#> 55  kkkpf4171         outp     2016-05-07              Glinifera 2016-05-07
#> 56  kkkpf4171         outp     2016-05-07          beta-blockers 2016-05-07
#> 57  kkkpf4171         outp     2016-05-07         Anticoagulants 2016-05-07
#> 58  opzoz2779         Hosp     2016-09-06         Anticoagulants 2016-09-06
#> 59  opzoz2779         Hosp     2016-09-06 antiplatelet medicines 2016-09-06
#> 60  opzoz2779         Hosp     2016-09-06                insulin 2016-09-06
#> 61  opzoz2779         Hosp     2016-09-06          Sulfonylureas 2016-09-06
#> 62  hwjjh4986         outp     2016-09-18             Biguanides 2016-09-18
#> 63  hwjjh4986         outp     2016-09-18              Glinifera 2016-09-18
#> 64  hwjjh4986         outp     2016-09-18                insulin 2016-09-18
#> 65  hwjjh4986         outp     2016-09-18 antiplatelet medicines 2016-09-18
#> 66  hwjjh4986         outp     2016-09-18                   ARBs 2016-09-18
#> 67  hwjjh4986         outp     2016-09-18                Statins 2016-09-18
#> 68  cpqzj5218         Hosp     2017-02-01         Anticoagulants 2017-02-01
#> 69  cpqzj5218         Hosp     2017-02-01 Lipid-regulating drugs 2017-02-01
#> 70  cpqzj5218         Hosp     2017-02-01 antiplatelet medicines 2017-02-01
#> 71  cpqzj5218         Hosp     2017-02-01                Statins 2017-02-01
#> 72  cpqzj5218         Hosp     2017-02-01                insulin 2017-02-01
#> 73  cpqzj5218         Hosp     2017-02-01                   ARBs 2017-02-01
#> 74  cpqzj5218         Hosp     2017-02-01             Biguanides 2017-02-01
#> 75  cpqzj5218         Hosp     2017-02-01          beta-blockers 2017-02-01
#> 76  cpqzj5218         Hosp     2017-02-01              Glinifera 2017-02-01
#> 77  oiegk5170         Hosp     2017-02-16          Sulfonylureas 2017-02-16
#> 78  oiegk5170         Hosp     2017-02-16                insulin 2017-02-16
#> 79  oiegk5170         Hosp     2017-02-16                Statins 2017-02-16
#> 80  oiegk5170         Hosp     2017-02-16             Biguanides 2017-02-16
#> 81  oiegk5170         Hosp     2017-02-16              Glinifera 2017-02-16
#> 82  oiegk5170         Hosp     2017-02-16         Anticoagulants 2017-02-16
#> 83  oiegk5170         Hosp     2017-02-16 antiplatelet medicines 2017-02-16
#> 84  oiegk5170         Hosp     2017-02-16          beta-blockers 2017-02-16
#> 85  oiegk5170         Hosp     2017-02-16                   ARBs 2017-02-16
#> 86  pprzx3026         outp     2017-05-10 antiplatelet medicines 2017-05-10
#> 87  pprzx3026         outp     2017-05-10          beta-blockers 2017-05-10
#> 88  pprzx3026         outp     2017-05-10 Lipid-regulating drugs 2017-05-10
#> 89  pprzx3026         outp     2017-05-10              Glinifera 2017-05-10
#> 90  pprzx3026         outp     2017-05-10          Sulfonylureas 2017-05-10
#> 91  pprzx3026         outp     2017-05-10         Anticoagulants 2017-05-10
#> 92  pprzx3026         outp     2017-05-10                insulin 2017-05-10
#> 93  pprzx3026         outp     2017-05-10             Biguanides 2017-05-10
#> 94  pprzx3026         outp     2017-05-10                   ARBs 2017-05-10
#> 95  wuukc4654         outp     2017-08-01 antiplatelet medicines 2017-08-01
#> 96  wuukc4654         outp     2017-08-01                insulin 2017-08-01
#> 97  wuukc4654         outp     2017-08-01 Lipid-regulating drugs 2017-08-01
#> 98  wuukc4654         outp     2017-08-01          beta-blockers 2017-08-01
#> 99  wuukc4654         outp     2017-08-01              Glinifera 2017-08-01
#> 100 vlfya2039         outp     2017-12-05          Sulfonylureas 2017-12-05
#> 101 vlfya2039         outp     2017-12-05                Statins 2017-12-05
#> 102 vlfya2039         outp     2017-12-05         Anticoagulants 2017-12-05
#> 103 vlfya2039         outp     2017-12-05                   ARBs 2017-12-05
#> 104 vlfya2039         outp     2017-12-05             Biguanides 2017-12-05
#> 105 vlfya2039         outp     2017-12-05              Glinifera 2017-12-05
#> 106 vlfya2039         outp     2017-12-05 Lipid-regulating drugs 2017-12-05
#> 107 vlfya2039         outp     2017-12-05          beta-blockers 2017-12-05
#> 108 vlfya2039         outp     2017-12-05 antiplatelet medicines 2017-12-05
#> 109 vlfya2039         outp     2017-12-05                insulin 2017-12-05
#> 110 fwctd1785         outp     2017-12-06          Sulfonylureas 2017-12-06
#> 111 fwctd1785         outp     2017-12-06         Anticoagulants 2017-12-06
#> 112 fwctd1785         outp     2017-12-06              Glinifera 2017-12-06
#> 113 fwctd1785         outp     2017-12-06                Statins 2017-12-06
#> 114 gbbug3812         Hosp     2017-12-23          beta-blockers 2017-12-23
#> 115 gbbug3812         Hosp     2017-12-23              Glinifera 2017-12-23
#> 116 gbbug3812         Hosp     2017-12-23          Sulfonylureas 2017-12-23
#> 117 gbbug3812         Hosp     2017-12-23                insulin 2017-12-23
#> 118 gbbug3812         Hosp     2017-12-23 Lipid-regulating drugs 2017-12-23
#> 119 gbbug3812         Hosp     2017-12-23         Anticoagulants 2017-12-23
#> 120 cpttc5403         Hosp     2018-02-06                Statins 2018-02-06
#> 121 cpttc5403         Hosp     2018-02-06          beta-blockers 2018-02-06
#> 122 cpttc5403         Hosp     2018-02-06 Lipid-regulating drugs 2018-02-06
#> 123 cpttc5403         Hosp     2018-02-06                insulin 2018-02-06
#> 124 cpttc5403         Hosp     2018-02-06 antiplatelet medicines 2018-02-06
#> 125 cpttc5403         Hosp     2018-02-06          Sulfonylureas 2018-02-06
#> 126 vfpeh2440         outp     2018-06-22 antiplatelet medicines 2018-06-22
#> 127 vfpeh2440         outp     2018-06-22                   ARBs 2018-06-22
#> 128 vfpeh2440         outp     2018-06-22          beta-blockers 2018-06-22
#> 129 vfpeh2440         outp     2018-06-22             Biguanides 2018-06-22
#> 130 vfpeh2440         outp     2018-06-22          Sulfonylureas 2018-06-22
#> 131 vfpeh2440         outp     2018-06-22                insulin 2018-06-22
#> 132 vfpeh2440         outp     2018-06-22         Anticoagulants 2018-06-22
#> 133 vfpeh2440         outp     2018-06-22              Glinifera 2018-06-22
#> 134 vfpeh2440         outp     2018-06-22 Lipid-regulating drugs 2018-06-22
#> 135 vfpeh2440         outp     2018-06-22                Statins 2018-06-22
#> 136 qgauj2201         outp     2018-07-31                insulin 2018-07-31
#> 137 qgauj2201         outp     2018-07-31                   ARBs 2018-07-31
#> 138 qgauj2201         outp     2018-07-31 antiplatelet medicines 2018-07-31
#> 139 qgauj2201         outp     2018-07-31          Sulfonylureas 2018-07-31
#> 140 qgauj2201         outp     2018-07-31         Anticoagulants 2018-07-31
#> 141 qgauj2201         outp     2018-07-31              Glinifera 2018-07-31
#> 142 qgauj2201         outp     2018-07-31 Lipid-regulating drugs 2018-07-31
#> 143 qgauj2201         outp     2018-07-31          beta-blockers 2018-07-31
#> 144 fnlaz3245         Hosp     2018-08-09                   ARBs 2018-08-09
#> 145 fnlaz3245         Hosp     2018-08-09 Lipid-regulating drugs 2018-08-09
#> 146 fnlaz3245         Hosp     2018-08-09          Sulfonylureas 2018-08-09
#> 147 fnlaz3245         Hosp     2018-08-09          beta-blockers 2018-08-09
#> 148 fnlaz3245         Hosp     2018-08-09              Glinifera 2018-08-09
#> 149 fnlaz3245         Hosp     2018-08-09         Anticoagulants 2018-08-09
#> 150 fnlaz3245         Hosp     2018-08-09 antiplatelet medicines 2018-08-09
#> 151 fnlaz3245         Hosp     2018-08-09                insulin 2018-08-09
#> 152 xqnri2590         Hosp     2018-08-29                   ARBs 2018-08-29
#> 153 xqnri2590         Hosp     2018-08-29         Anticoagulants 2018-08-29
#> 154 xqnri2590         Hosp     2018-08-29 Lipid-regulating drugs 2018-08-29
#> 155 xqnri2590         Hosp     2018-08-29          beta-blockers 2018-08-29
#> 156 qdnjd5149         Hosp     2018-09-12          Sulfonylureas 2018-09-12
#> 157 qdnjd5149         Hosp     2018-09-12         Anticoagulants 2018-09-12
#> 158 qdnjd5149         Hosp     2018-09-12 antiplatelet medicines 2018-09-12
#> 159 qdnjd5149         Hosp     2018-09-12                Statins 2018-09-12
#> 160 rviko4185         outp     2018-10-04                   ARBs 2018-10-04
#> 161 rviko4185         outp     2018-10-04 antiplatelet medicines 2018-10-04
#> 162 rviko4185         outp     2018-10-04                insulin 2018-10-04
#> 163 tmcsy4261         Hosp     2018-10-07                insulin 2018-10-07
#> 164 tmcsy4261         Hosp     2018-10-07                   ARBs 2018-10-07
#> 165 tmcsy4261         Hosp     2018-10-07 Lipid-regulating drugs 2018-10-07
#> 166 tmcsy4261         Hosp     2018-10-07 antiplatelet medicines 2018-10-07
#> 167 tmcsy4261         Hosp     2018-10-07          beta-blockers 2018-10-07
#> 168 ikveg4786         Hosp     2018-12-12 antiplatelet medicines 2018-12-12
#> 169 ikveg4786         Hosp     2018-12-12             Biguanides 2018-12-12
#> 170 ikveg4786         Hosp     2018-12-12                Statins 2018-12-12
#> 171 ikveg4786         Hosp     2018-12-12         Anticoagulants 2018-12-12
#> 172 ikveg4786         Hosp     2018-12-12                insulin 2018-12-12
#> 173 sxklc2322         Hosp     2019-01-05             Biguanides 2019-01-05
#> 174 sxklc2322         Hosp     2019-01-05              Glinifera 2019-01-05
#> 175 sxklc2322         Hosp     2019-01-05 Lipid-regulating drugs 2019-01-05
#> 176 sxklc2322         Hosp     2019-01-05          beta-blockers 2019-01-05
#> 177 sxklc2322         Hosp     2019-01-05                Statins 2019-01-05
#> 178 sxklc2322         Hosp     2019-01-05 antiplatelet medicines 2019-01-05
#> 179 sxklc2322         Hosp     2019-01-05                   ARBs 2019-01-05
#> 180 sxklc2322         Hosp     2019-01-05         Anticoagulants 2019-01-05
#> 181 sxklc2322         Hosp     2019-01-05          Sulfonylureas 2019-01-05
#> 182 sxklc2322         Hosp     2019-01-05                insulin 2019-01-05
#> 183 hnfot2199         Hosp     2019-02-16          Sulfonylureas 2019-02-16
#> 184 hnfot2199         Hosp     2019-02-16              Glinifera 2019-02-16
#> 185 hnfot2199         Hosp     2019-02-16                   ARBs 2019-02-16
#> 186 hnfot2199         Hosp     2019-02-16          beta-blockers 2019-02-16
#> 187 hnfot2199         Hosp     2019-02-16 antiplatelet medicines 2019-02-16
#> 188 hnfot2199         Hosp     2019-02-16 Lipid-regulating drugs 2019-02-16
#> 189 hnfot2199         Hosp     2019-02-16         Anticoagulants 2019-02-16
#> 190 hnfot2199         Hosp     2019-02-16                insulin 2019-02-16
#> 191 hnfot2199         Hosp     2019-02-16             Biguanides 2019-02-16
#> 192 crght3835         outp     2019-04-18          beta-blockers 2019-04-18
#> 193 crght3835         outp     2019-04-18          Sulfonylureas 2019-04-18
#> 194 crght3835         outp     2019-04-18              Glinifera 2019-04-18
#> 195 crght3835         outp     2019-04-18                Statins 2019-04-18
#> 196 crght3835         outp     2019-04-18 Lipid-regulating drugs 2019-04-18
#> 197 crght3835         outp     2019-04-18 antiplatelet medicines 2019-04-18
#> 198 crght3835         outp     2019-04-18             Biguanides 2019-04-18
#> 199 rlhki2360         outp     2020-04-16         Anticoagulants 2020-04-16
#> 200 rlhki2360         outp     2020-04-16                   ARBs 2020-04-16
#> 201 rlhki2360         outp     2020-04-16 Lipid-regulating drugs 2020-04-16
#> 202 rlhki2360         outp     2020-04-16          beta-blockers 2020-04-16
#> 203 rlhki2360         outp     2020-04-16          Sulfonylureas 2020-04-16
#> 204 oavnw2458         outp     2020-06-08              Glinifera 2020-06-08
#> 205 oavnw2458         outp     2020-06-08                insulin 2020-06-08
#> 206 oavnw2458         outp     2020-06-08 antiplatelet medicines 2020-06-08
#> 207 oavnw2458         outp     2020-06-08                Statins 2020-06-08
#> 208 oavnw2458         outp     2020-06-08 Lipid-regulating drugs 2020-06-08
#> 209 oavnw2458         outp     2020-06-08         Anticoagulants 2020-06-08
#> 210 oavnw2458         outp     2020-06-08          Sulfonylureas 2020-06-08
#> 
#> $lab
#>      visit_id patient_type admission_date lab_name lab_va  test_date
#> 1   zgvbb2931         Hosp     2014-05-14      HDL      2 2014-05-14
#> 2   zgvbb2931         Hosp     2014-05-14       TC      5 2014-05-14
#> 3   zgvbb2931         Hosp     2014-05-14    HbA1c     11 2014-05-14
#> 4   zgvbb2931         Hosp     2014-05-14      WBC      1 2014-05-14
#> 5   egclx5067         Hosp     2014-06-23    HbA1c     10 2014-06-23
#> 6   egclx5067         Hosp     2014-06-23      HDL      5 2014-06-23
#> 7   egclx5067         Hosp     2014-06-23      WBC      4 2014-06-23
#> 8   tuhol1267         outp     2015-02-26       TG     20 2015-02-26
#> 9   tuhol1267         outp     2015-02-26      HDL      1 2015-02-26
#> 10  tuhol1267         outp     2015-02-26      LDL      1 2015-02-26
#> 11  yyiia3143         Hosp     2015-05-24      HDL      4 2015-05-24
#> 12  yyiia3143         Hosp     2015-05-24      WBC      6 2015-05-24
#> 13  yyiia3143         Hosp     2015-05-24       TC      6 2015-05-24
#> 14  yyiia3143         Hosp     2015-05-24       TG     13 2015-05-24
#> 15  wvvwk5077         outp     2015-07-08      LDL      3 2015-07-08
#> 16  wvvwk5077         outp     2015-07-08      HDL      1 2015-07-08
#> 17  wvvwk5077         outp     2015-07-08    HbA1c      8 2015-07-08
#> 18  wvvwk5077         outp     2015-07-08      WBC     19 2015-07-08
#> 19  wvvwk5077         outp     2015-07-08       TG     10 2015-07-08
#> 20  qimcg2445         Hosp     2015-08-08       TG     12 2015-08-08
#> 21  qimcg2445         Hosp     2015-08-08       TC      5 2015-08-08
#> 22  qimcg2445         Hosp     2015-08-08      LDL      2 2015-08-08
#> 23  qimcg2445         Hosp     2015-08-08      HDL      2 2015-08-08
#> 24  tnqqj2754         Hosp     2015-11-12      WBC     11 2015-11-12
#> 25  tnqqj2754         Hosp     2015-11-12      HDL      5 2015-11-12
#> 26  tnqqj2754         Hosp     2015-11-12       TG     13 2015-11-12
#> 27  tnqqj2754         Hosp     2015-11-12      LDL      4 2015-11-12
#> 28  odgxj5634         Hosp     2016-02-11      WBC     11 2016-02-11
#> 29  odgxj5634         Hosp     2016-02-11      HDL      2 2016-02-11
#> 30  odgxj5634         Hosp     2016-02-11      LDL      3 2016-02-11
#> 31  odgxj5634         Hosp     2016-02-11    HbA1c      4 2016-02-11
#> 32  odgxj5634         Hosp     2016-02-11       TG      2 2016-02-11
#> 33  odgxj5634         Hosp     2016-02-11       TC      1 2016-02-11
#> 34  kkkpf4171         outp     2016-05-07       TG      2 2016-05-07
#> 35  kkkpf4171         outp     2016-05-07      LDL      2 2016-05-07
#> 36  kkkpf4171         outp     2016-05-07    HbA1c      9 2016-05-07
#> 37  kkkpf4171         outp     2016-05-07      WBC     13 2016-05-07
#> 38  opzoz2779         Hosp     2016-09-06      WBC     16 2016-09-06
#> 39  opzoz2779         Hosp     2016-09-06       TC      6 2016-09-06
#> 40  opzoz2779         Hosp     2016-09-06       TG      6 2016-09-06
#> 41  opzoz2779         Hosp     2016-09-06      HDL      4 2016-09-06
#> 42  opzoz2779         Hosp     2016-09-06      LDL      1 2016-09-06
#> 43  opzoz2779         Hosp     2016-09-06    HbA1c      6 2016-09-06
#> 44  hwjjh4986         outp     2016-09-18    HbA1c      3 2016-09-18
#> 45  hwjjh4986         outp     2016-09-18      WBC      4 2016-09-18
#> 46  hwjjh4986         outp     2016-09-18       TG     20 2016-09-18
#> 47  hwjjh4986         outp     2016-09-18      HDL      5 2016-09-18
#> 48  hwjjh4986         outp     2016-09-18       TC      2 2016-09-18
#> 49  cpqzj5218         Hosp     2017-02-01       TC      5 2017-02-01
#> 50  cpqzj5218         Hosp     2017-02-01      LDL      1 2017-02-01
#> 51  cpqzj5218         Hosp     2017-02-01       TG     20 2017-02-01
#> 52  cpqzj5218         Hosp     2017-02-01      HDL      4 2017-02-01
#> 53  cpqzj5218         Hosp     2017-02-01    HbA1c      8 2017-02-01
#> 54  cpqzj5218         Hosp     2017-02-01      WBC      5 2017-02-01
#> 55  oiegk5170         Hosp     2017-02-16      WBC      4 2017-02-16
#> 56  oiegk5170         Hosp     2017-02-16    HbA1c      9 2017-02-16
#> 57  oiegk5170         Hosp     2017-02-16      HDL      5 2017-02-16
#> 58  oiegk5170         Hosp     2017-02-16       TC      5 2017-02-16
#> 59  pprzx3026         outp     2017-05-10      WBC     17 2017-05-10
#> 60  pprzx3026         outp     2017-05-10       TC      4 2017-05-10
#> 61  pprzx3026         outp     2017-05-10    HbA1c      5 2017-05-10
#> 62  pprzx3026         outp     2017-05-10      HDL      4 2017-05-10
#> 63  pprzx3026         outp     2017-05-10      LDL      4 2017-05-10
#> 64  wuukc4654         outp     2017-08-01       TC      3 2017-08-01
#> 65  wuukc4654         outp     2017-08-01      LDL      5 2017-08-01
#> 66  wuukc4654         outp     2017-08-01      HDL      3 2017-08-01
#> 67  wuukc4654         outp     2017-08-01    HbA1c      9 2017-08-01
#> 68  wuukc4654         outp     2017-08-01      WBC     11 2017-08-01
#> 69  wuukc4654         outp     2017-08-01       TG      6 2017-08-01
#> 70  vlfya2039         outp     2017-12-05       TG      3 2017-12-05
#> 71  vlfya2039         outp     2017-12-05       TC      3 2017-12-05
#> 72  vlfya2039         outp     2017-12-05      LDL      4 2017-12-05
#> 73  vlfya2039         outp     2017-12-05    HbA1c      8 2017-12-05
#> 74  fwctd1785         outp     2017-12-06    HbA1c      8 2017-12-06
#> 75  fwctd1785         outp     2017-12-06       TG      8 2017-12-06
#> 76  fwctd1785         outp     2017-12-06      WBC     18 2017-12-06
#> 77  fwctd1785         outp     2017-12-06      LDL      5 2017-12-06
#> 78  fwctd1785         outp     2017-12-06      HDL      5 2017-12-06
#> 79  fwctd1785         outp     2017-12-06       TC      3 2017-12-06
#> 80  gbbug3812         Hosp     2017-12-23      LDL      3 2017-12-23
#> 81  gbbug3812         Hosp     2017-12-23      HDL      3 2017-12-23
#> 82  gbbug3812         Hosp     2017-12-23    HbA1c     12 2017-12-23
#> 83  gbbug3812         Hosp     2017-12-23       TC      3 2017-12-23
#> 84  gbbug3812         Hosp     2017-12-23      WBC     13 2017-12-23
#> 85  cpttc5403         Hosp     2018-02-06       TG      6 2018-02-06
#> 86  cpttc5403         Hosp     2018-02-06      WBC      2 2018-02-06
#> 87  cpttc5403         Hosp     2018-02-06    HbA1c      4 2018-02-06
#> 88  cpttc5403         Hosp     2018-02-06      HDL      5 2018-02-06
#> 89  cpttc5403         Hosp     2018-02-06       TC      4 2018-02-06
#> 90  vfpeh2440         outp     2018-06-22       TG      7 2018-06-22
#> 91  vfpeh2440         outp     2018-06-22      WBC     20 2018-06-22
#> 92  vfpeh2440         outp     2018-06-22       TC      4 2018-06-22
#> 93  vfpeh2440         outp     2018-06-22    HbA1c     10 2018-06-22
#> 94  vfpeh2440         outp     2018-06-22      LDL      5 2018-06-22
#> 95  qgauj2201         outp     2018-07-31    HbA1c      2 2018-07-31
#> 96  qgauj2201         outp     2018-07-31       TC      5 2018-07-31
#> 97  qgauj2201         outp     2018-07-31       TG     14 2018-07-31
#> 98  fnlaz3245         Hosp     2018-08-09       TC      5 2018-08-09
#> 99  fnlaz3245         Hosp     2018-08-09    HbA1c      9 2018-08-09
#> 100 fnlaz3245         Hosp     2018-08-09       TG     13 2018-08-09
#> 101 xqnri2590         Hosp     2018-08-29      WBC     10 2018-08-29
#> 102 xqnri2590         Hosp     2018-08-29    HbA1c     10 2018-08-29
#> 103 xqnri2590         Hosp     2018-08-29      HDL      3 2018-08-29
#> 104 xqnri2590         Hosp     2018-08-29      LDL      5 2018-08-29
#> 105 qdnjd5149         Hosp     2018-09-12      LDL      1 2018-09-12
#> 106 qdnjd5149         Hosp     2018-09-12    HbA1c      9 2018-09-12
#> 107 qdnjd5149         Hosp     2018-09-12       TC      7 2018-09-12
#> 108 qdnjd5149         Hosp     2018-09-12       TG      7 2018-09-12
#> 109 rviko4185         outp     2018-10-04    HbA1c      4 2018-10-04
#> 110 rviko4185         outp     2018-10-04      LDL      5 2018-10-04
#> 111 rviko4185         outp     2018-10-04      HDL      3 2018-10-04
#> 112 rviko4185         outp     2018-10-04       TC      1 2018-10-04
#> 113 rviko4185         outp     2018-10-04       TG     19 2018-10-04
#> 114 rviko4185         outp     2018-10-04      WBC     19 2018-10-04
#> 115 tmcsy4261         Hosp     2018-10-07      LDL      5 2018-10-07
#> 116 tmcsy4261         Hosp     2018-10-07      WBC     16 2018-10-07
#> 117 tmcsy4261         Hosp     2018-10-07      HDL      2 2018-10-07
#> 118 tmcsy4261         Hosp     2018-10-07    HbA1c      5 2018-10-07
#> 119 tmcsy4261         Hosp     2018-10-07       TG      7 2018-10-07
#> 120 ikveg4786         Hosp     2018-12-12      HDL      5 2018-12-12
#> 121 ikveg4786         Hosp     2018-12-12      LDL      4 2018-12-12
#> 122 ikveg4786         Hosp     2018-12-12      WBC     10 2018-12-12
#> 123 ikveg4786         Hosp     2018-12-12       TC      5 2018-12-12
#> 124 ikveg4786         Hosp     2018-12-12       TG     20 2018-12-12
#> 125 sxklc2322         Hosp     2019-01-05    HbA1c      8 2019-01-05
#> 126 sxklc2322         Hosp     2019-01-05       TC      2 2019-01-05
#> 127 sxklc2322         Hosp     2019-01-05      LDL      2 2019-01-05
#> 128 hnfot2199         Hosp     2019-02-16       TC      1 2019-02-16
#> 129 hnfot2199         Hosp     2019-02-16    HbA1c     12 2019-02-16
#> 130 hnfot2199         Hosp     2019-02-16      HDL      4 2019-02-16
#> 131 hnfot2199         Hosp     2019-02-16      LDL      5 2019-02-16
#> 132 crght3835         outp     2019-04-18       TC      7 2019-04-18
#> 133 crght3835         outp     2019-04-18    HbA1c     11 2019-04-18
#> 134 crght3835         outp     2019-04-18      WBC      9 2019-04-18
#> 135 rlhki2360         outp     2020-04-16       TC      4 2020-04-16
#> 136 rlhki2360         outp     2020-04-16       TG      9 2020-04-16
#> 137 rlhki2360         outp     2020-04-16      WBC     12 2020-04-16
#> 138 oavnw2458         outp     2020-06-08       TC      1 2020-06-08
#> 139 oavnw2458         outp     2020-06-08      HDL      3 2020-06-08
#> 140 oavnw2458         outp     2020-06-08       TG     12 2020-06-08
#> 141 oavnw2458         outp     2020-06-08      LDL      2 2020-06-08
#> 142 oavnw2458         outp     2020-06-08      WBC      3 2020-06-08
#> 143 oavnw2458         outp     2020-06-08    HbA1c     11 2020-06-08
#> 
#> $diag
#> # A tibble: 32 × 5
#>    visit_id  patient_type admission_date diag_date  diag_union                  
#>    <chr>     <chr>        <date>         <date>     <chr>                       
#>  1 zgvbb2931 Hosp         2014-05-14     2014-05-15 Dyslipidemia,high blood pre…
#>  2 egclx5067 Hosp         2014-06-23     2014-06-24 diabetes,high blood pressur…
#>  3 tuhol1267 outp         2015-02-26     2015-02-27 high blood pressure,NA,Stro…
#>  4 yyiia3143 Hosp         2015-05-24     2015-05-25 diabetes,NA,high blood pres…
#>  5 wvvwk5077 outp         2015-07-08     2015-07-09 Stroke,Mild liver disease,M…
#>  6 qimcg2445 Hosp         2015-08-08     2015-08-09 high blood pressure,Connect…
#>  7 tnqqj2754 Hosp         2015-11-12     2015-11-13 Symptomatic neuropathy,Dysl…
#>  8 odgxj5634 Hosp         2016-02-11     2016-02-12 diabetes,NA,Peptic ulcers,M…
#>  9 kkkpf4171 outp         2016-05-07     2016-05-08 Dyslipidemia,Chronic lung d…
#> 10 opzoz2779 Hosp         2016-09-06     2016-09-07 Chronic lung disease,Connec…
#> # … with 22 more rows
#>